==> Scripts/2c_Adding_metadata_KRD2022.Rmd <==
---
title: "Additional metadata - KRD 2022 livestock data"
author: "Beatrice Cornu Hewitt" 
date: "`r Sys.time()`"
---

Some metadata needs to be added which was not originally added when creating the phyloseq object, e.g. the KRD 2022 livestock data
We add this now to the ps_complete phyloseq object

```{r}
ps_complete_original <- readRDS("../Output/Phyloseq/x superseded/ps_complete.rds")
original_samdata <- ps_complete_original@sam_data
original_samdata_df <- data.frame(original_samdata)

# Load the KRD livestock characteristics
KRD2022_livestock_exp <- read.csv("../Input/Livestock_computations/KRD2022_livestock_characteristics.csv")

# Change the name of column ID to VGO3_ID to match with VGO3_metadata_all
KRD2022_livestock_exp <- KRD2022_livestock_exp %>%
  rename(VGO3_ID = respnr)
# Convert to character var to match with original_samdata_df
KRD2022_livestock_exp$VGO3_ID <- as.character(KRD2022_livestock_exp$VGO3_ID)

new_metadata_KRD2022 <- left_join(original_samdata_df, KRD2022_livestock_exp, by = "VGO3_ID")
rownames(new_metadata_KRD2022) <- rownames(original_samdata_df)
new_sample_data <- sample_data(new_metadata_KRD2022)

ps_new <- ps_complete_original
ps_new@sam_data <- new_sample_data

saveRDS(ps_new, "../Output/Phyloseq/ps_complete.rds")
```
