==> Scripts/6_Alpha_diversity.Rmd <==
---
title: "Alpha diversity"
author: "Beatrice Cornu Hewitt" 
date: "`r Sys.time()`"
---

# Packages
```{r}
library(phyloseq); library(microbiomer); library(ggplot2); library(cowplot); library(patchwork); library(scales); library(tidyverse); library(car); library(lmtest); library(ggsignif)

```
# Residential livestock exposure
## Data
```{r}
ps <- readRDS("../Output/Phyloseq/2_ps_complete.Rds")
```
## Prepare data
```{r}
ps_CP <- ps 
ps_cp <- subset_samples(ps, population == "CP")
# Extract the sample data from the phyloseq object
sample_data_df <- data.frame(sample_data(ps_cp))

# Calculate Pearson correlation
cor.test(sample_data_df$age, sample_data_df$KRD_nAnyFarmWghtDist_3000m) # p-value = 0.003351
cor.test(sample_data_df$age, sample_data_df$KRD_nGoatFarmWghtDist.3000m.sum) # p-value = 0.2427
cor.test(sample_data_df$age, sample_data_df$KRD_nGoatsWghtDist.3000m.sum) # p-value = 0.05264
cor.test(sample_data_df$age, sample_data_df$KRD_MindistGoatFarm50.NEG) # p-value = 0.6065

# NP 
# Remove low abundant ASVs and convert counts to relative abundances
ps_np_filtered <- subset_samples(ps, niche == "NP") %>% pres_abund_filter() %>% to_RA()
# A total of 1527 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 19079 ASVs excluded).

# Subset for CP only
ps_np_cp <- subset_samples(ps_np_filtered, population == "CP")
otu_np_cp <- as.data.frame(otu_table(ps_np_cp)) %>% t()
meta_np_cp <- meta_to_df(ps_np_cp)

# OP 
# Remove low abundant ASVs and convert counts to relative abundances
ps_op_filtered <- subset_samples(ps, niche == "OP") %>% pres_abund_filter() %>% to_RA()
# A total of 1346 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 19260 ASVs excluded).

# Subset for CP only
ps_op_cp <- subset_samples(ps_op_filtered, population == "CP")
otu_op_cp <- as.data.frame(otu_table(ps_op_cp)) %>% t()
meta_op_cp <- meta_to_df(ps_op_cp)
```
# NP vs OP for the residents 
```{r}
# Subset data for CP population
# Extract metadata
meta_data_cp <- meta_to_df(ps_cp) %>% 
  rownames_to_column(var = "SampleID")

# Shapiro wilk test for normality
shapiro.test(meta_data_cp$Shannon[meta_data_cp$niche == "NP"]) 
shapiro.test(meta_data_cp$Shannon[meta_data_cp$niche == "OP"])
shapiro.test(meta_data_cp$Observed[meta_data_cp$niche == "NP"]) 
shapiro.test(meta_data_cp$Observed[meta_data_cp$niche == "OP"]) 
shapiro.test(meta_data_cp$qPCR[meta_data_cp$niche == "NP"]) 
shapiro.test(meta_data_cp$qPCR[meta_data_cp$niche == "OP"]) 

# All are significant therefore shannon, observed and qPCR are not normally distributed in the the NP or OP
hist(meta_data_cp$Shannon[meta_data_cp$niche == "NP"])
hist(meta_data_cp$Shannon[meta_data_cp$niche == "OP"])
hist(meta_data_cp$Observed[meta_data_cp$niche == "NP"])
hist(meta_data_cp$Observed[meta_data_cp$niche == "OP"])
hist(meta_data_cp$qPCR[meta_data_cp$niche == "NP"])
hist(meta_data_cp$qPCR[meta_data_cp$niche == "OP"])

# Use Wilcoxon Rank-Sum Test (Non-Parametric Alternative to t-test) to test for differenes between OP and NP
shannon_np_op <- wilcox.test(Shannon ~ niche, data = meta_data_cp)
observed_np_op <- wilcox.test(Observed ~ niche, data = meta_data_cp)
qPCR_np_op <- wilcox.test(qPCR ~ niche, data = meta_data_cp)

# Extract p-values 
shannon_p <- format(shannon_np_op$p.value, scientific = TRUE, digits = 3)
observed_p <- format(observed_np_op$p.value, scientific = TRUE, digits = 3)
qPCR_p <- format(qPCR_np_op$p.value, scientific = TRUE, digits = 3)

# Create a data frame to store the results
NPvsOP_diversity_res <- data.frame(
  Test = c("Shannon", "Observed", "qPCR"),
  W_Statistic = c(shannon_np_op$statistic, observed_np_op$statistic, qPCR_np_op$statistic),
  P_Value = c(shannon_p, observed_p, qPCR_p)
)

write.csv(NPvsOP_diversity_res, "../Output/Alpha_diversity/NPvsOP_shann_obs_qPCR.csv")

# Format p-values, and label as "<0.001" if they are smaller than 0.001
shannon_p <- ifelse(shannon_np_op$p.value < 0.001, "<0.001", format(shannon_np_op$p.value, scientific = TRUE, digits = 3))
observed_p <- ifelse(observed_np_op$p.value < 0.001, "<0.001", format(observed_np_op$p.value, scientific = TRUE, digits = 3))
qPCR_p <- ifelse(qPCR_np_op$p.value < 0.001, "<0.001", format(qPCR_np_op$p.value, scientific = TRUE, digits = 3))


# Define custom colors for the niches
box_colors <- c("NP" = "#dd968e", "OP" = "#547484")
scatter_colors <- c("NP" = muted("#dd968e", l = 40), "OP" = muted("#547484", l = 40))

# Plot Shannon diversity
plot_shannon <- ggplot(meta_data_cp, aes(x = niche, y = Shannon, fill = niche)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Add violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8) +  # Overlay boxplot
  labs(
    x = "Niche", 
    y = "Shannon Diversity"  ) +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    plot.title = element_text(size = 18, hjust = 0.5)
  ) +
  scale_fill_manual(values = box_colors) +  # Custom colors for violin and boxplots
  # Add p-value annotation
  annotate(
    "text",
    x = 1.5,  # Position between NP and OP
    y = max(meta_data_cp$Shannon, na.rm = TRUE) + 0.1,  # Slightly above the max value
    label = bquote(italic(p)~" value"*.(shannon_p)),
    size = 4
  )

# Plot Observed species diversity
plot_observed <- ggplot(meta_data_cp, aes(x = niche, y = Observed, fill = niche)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Add violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8) +  # Overlay boxplot
  labs(
    x = "Niche", 
    y = "Observed richness"  ) +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    plot.title = element_text(size = 18, hjust = 0.5)
  ) +
  scale_fill_manual(values = box_colors) +  # Custom colors for violin and boxplots
  # Add p-value annotation
  annotate(
    "text",
    x = 1.5,  # Position between NP and OP
    y = max(meta_data_cp$Observed, na.rm = TRUE) + 0.1,  # Slightly above the max value
    label = bquote(italic(p)~" value"*.(observed_p)),
    size = 4
  )

# Plot qPCR 
meta_data_cp$qPCR_log
plot_qPCR <- ggplot(meta_data_cp, aes(x = niche, y = qPCR_log, fill = niche)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Add violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8) +  # Overlay boxplot
  labs(
    x = "Niche", 
    y = "16S qPCR (log)") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    plot.title = element_text(size = 18, hjust = 0.5)
  ) +
  scale_fill_manual(values = box_colors) +  # Custom colors for violin and boxplots
  # Add p-value annotation
  annotate(
    "text",
    x = 1.5,  # Position between NP and OP
    y = max(meta_data_cp$qPCR_log, na.rm = TRUE) + 0.1,  # Slightly above the max value
    label = bquote(italic(p)~" value"*.(qPCR_p)),
    size = 4
  )

# Shannon figure for main manuscript
ggsave("../Output/Alpha_diversity/NP_vs_OP_violin_boxplot_shannon.svg", plot_shannon, width = 3, height = 5)

# Supplementary figure (observed richness and qPCR)
NPvsOP_diversity <- plot_grid(plot_shannon, plot_observed, plot_qPCR, nrow = 1, align = "hv")
NPvsOP_diversity

# Save the plots
ggsave("../Output/Alpha_diversity/NP_vs_OP_violin_boxplot_diversity.svg", NPvsOP_diversity, width = 8, height = 5)

```
# NP vs OP for the residents & goat farmers 
```{r}
# Subset data for CP population
# Extract metadata
meta_data_all <- meta_to_df(ps) %>% 
  rownames_to_column(var = "SampleID")

# Shapiro wilk test for normality
shapiro.test(meta_data_all$Shannon[meta_data_all$niche == "NP"]) 
shapiro.test(meta_data_all$Shannon[meta_data_all$niche == "OP"])
shapiro.test(meta_data_all$Observed[meta_data_all$niche == "NP"]) 
shapiro.test(meta_data_all$Observed[meta_data_all$niche == "OP"]) 
shapiro.test(meta_data_all$qPCR[meta_data_all$niche == "NP"]) 
shapiro.test(meta_data_all$qPCR[meta_data_all$niche == "OP"]) 

# All are significant therefore shannon, observed and qPCR are not normally distributed in the the NP or OP
hist(meta_data_all$Shannon[meta_data_all$niche == "NP"])
hist(meta_data_all$Shannon[meta_data_all$niche == "OP"])
hist(meta_data_all$Observed[meta_data_all$niche == "NP"])
hist(meta_data_all$Observed[meta_data_all$niche == "OP"])
hist(meta_data_all$qPCR[meta_data_all$niche == "NP"])
hist(meta_data_all$qPCR[meta_data_all$niche == "OP"])

# Use Wilcoxon Rank-Sum Test (Non-Parametric Alternative to t-test) to test for differenes between OP and NP
shannon_np_op_all <- wilcox.test(Shannon ~ niche, data = meta_data_all)
observed_np_op_all <- wilcox.test(Observed ~ niche, data = meta_data_all)
qPCR_np_op_all <- wilcox.test(qPCR ~ niche, data = meta_data_all)

# Extract p-values 
shannon_p_all <- format(shannon_np_op_all$p.value, scientific = TRUE, digits = 3)
observed_p_all <- format(observed_np_op_all$p.value, scientific = TRUE, digits = 3)
qPCR_p_all <- format(qPCR_np_op_all$p.value, scientific = TRUE, digits = 3)

# Create a data frame to store the results
NPvsOP_diversity_all <- data.frame(
  Test = c("Shannon", "Observed", "qPCR"),
  W_Statistic = c(shannon_np_op_all$statistic, observed_np_op_all$statistic, qPCR_np_op_all$statistic),
  P_Value = c(shannon_p_all, observed_p_all, qPCR_p_all)
)

write.csv(NPvsOP_diversity_all, "../Output/Alpha_diversity/NPvsOP_shann_obs_qPCR_all.csv")

# Format p-values, and label as "<0.001" if they are smaller than 0.001
shannon_p_all <- ifelse(shannon_np_op_all$p.value < 0.001, "<0.001", format(shannon_np_op_all$p.value, scientific = TRUE, digits = 3))
observed_p_all <- ifelse(observed_np_op_all$p.value < 0.001, "<0.001", format(observed_np_op_all$p.value, scientific = TRUE, digits = 3))
qPCR_p_all <- ifelse(qPCR_np_op_all$p.value < 0.001, "<0.001", format(qPCR_np_op_all$p.value, scientific = TRUE, digits = 3))


# Define custom colors for the niches
box_colors <- c("NP" = "#dd968e", "OP" = "#547484")
scatter_colors <- c("NP" = muted("#dd968e", l = 40), "OP" = muted("#547484", l = 40))

# Plot Shannon diversity
plot_shannon_all <- ggplot(meta_data_all, aes(x = niche, y = Shannon, fill = niche)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Add violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8) +  # Overlay boxplot
  labs(
    x = "Niche", 
    y = "Shannon diversity"  ) +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    plot.title = element_text(size = 18, hjust = 0.5)
  ) +
  scale_fill_manual(values = box_colors) +  # Custom colors for violin and boxplots
  # Add p-value annotation
  annotate(
    "text",
    x = 1.5,  # Position between NP and OP
    y = max(meta_data_all$Shannon, na.rm = TRUE) + 0.1,  # Slightly above the max value
    label = bquote(italic(p)~" value"*.(shannon_p_all)),
    size = 4
  )

# Plot Observed species diversity
plot_observed_all <- ggplot(meta_data_all, aes(x = niche, y = Observed, fill = niche)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Add violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8) +  # Overlay boxplot
  labs(
    x = "Niche", 
    y = "Observed richness"  ) +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    plot.title = element_text(size = 18, hjust = 0.5)
  ) +
  scale_fill_manual(values = box_colors) +  # Custom colors for violin and boxplots
  # Add p-value annotation
  annotate(
    "text",
    x = 1.5,  # Position between NP and OP
    y = max(meta_data_all$Observed, na.rm = TRUE) + 0.1,  # Slightly above the max value
    label = bquote(italic(p)~" value"*.(observed_p_all)),
    size = 4
  )

# Plot qPCR 
plot_qPCR_all <- ggplot(meta_data_all, aes(x = niche, y = qPCR_log, fill = niche)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Add violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8) +  # Overlay boxplot
  labs(
    x = "Niche", 
    y = "16S qPCR (log)") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    plot.title = element_text(size = 18, hjust = 0.5)
  ) +
  scale_fill_manual(values = box_colors) +  # Custom colors for violin and boxplots
  # Add p-value annotation
  annotate(
    "text",
    x = 1.5,  # Position between NP and OP
    y = max(meta_data_all$qPCR_log, na.rm = TRUE) + 0.1,  # Slightly above the max value
    label = bquote(italic(p)~" value"*.(qPCR_p_all)),
    size = 4
  )

# Shannon figure for main manuscript
NPvsOP_diversity_all <- plot_grid(plot_shannon_all, plot_observed_all, plot_qPCR_all, nrow = 1, align = "hv")
NPvsOP_diversity_all

# Save the plots
ggsave("../Output/Alpha_diversity/NP_vs_OP_violin_boxplot_diversity_all.svg", NPvsOP_diversity_all, width = 8, height = 5)

```

# Occupational livestock exposure 
## Data
```{r}
ps_cp_gf_matched <- readRDS("../Output/Phyloseq/3_ps_complete_matchedCPGF.rds")
```
## Prepare data
```{r}
# Extract the sample data from the phyloseq object
sample_data_cp_gf_df <- data.frame(sample_data(ps_cp_gf_matched))

# NP 
# Remove low abundant ASVs and convert counts to relative abundances
ps_cp_gf_np_filtered <- subset_samples(ps_cp_gf_matched, niche == "NP") %>% pres_abund_filter() %>% to_RA()
# A total of 732 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 8122 ASVs excluded).

# Extract data for linear models
otu_np_cp_gf <- as.data.frame(otu_table(ps_cp_gf_np_filtered)) %>% t()
meta_np_cp_gf <- meta_to_df(ps_cp_gf_np_filtered)

# OP 
# Remove low abundant ASVs and convert counts to relative abundances
ps_cp_gf_op_filtered <- subset_samples(ps_cp_gf_matched, niche == "OP") %>% pres_abund_filter() %>% to_RA()
# A total of 670 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 8184 ASVs excluded).

# Extract data for linear models
otu_op_cp_gf <- as.data.frame(otu_table(ps_cp_gf_op_filtered)) %>% t()
meta_op_cp_gf <- meta_to_df(ps_cp_gf_op_filtered)
```
## Linear models on Shannon and observed richness
```{r}
# Convert categorical variables to factors
str(meta_np_cp_gf$smoked_ever)
meta_np_cp_gf <- meta_np_cp_gf %>%
  mutate(
    population = as.factor(population),  # CP or GF
    gender = as.factor(gender),
    sampling_season = as.factor(sampling_season),
    smoked_ever = as.factor(smoked_ever)
  )

meta_op_cp_gf <- meta_op_cp_gf %>%
  mutate(
    population = as.factor(population), 
    gender = as.factor(gender),
    sampling_season = as.factor(sampling_season),
    smoked_ever = as.factor(smoked_ever)
  )

# Set CP as reference population
meta_np_cp_gf <- meta_np_cp_gf %>%
  mutate(population = factor(population, levels = c("CP", "GF")))  
meta_op_cp_gf <- meta_op_cp_gf %>%
  mutate(population = factor(population, levels = c("CP", "GF")))

# Linear regression for shannon 
lm_shannon_np_cp_gf <- lm(Shannon ~ population + age + sampling_season + gender + smoked_ever, data = meta_np_cp_gf) %>% 
  broom::tidy(., conf.int = T)
lm_shannon_np_cp_gf$livestock <- "Occupational exposure (GF vs CP)"
lm_shannon_np_cp_gf$niche <- "NP"

lm_shannon_op_cp_gf <- lm(Shannon ~ population + age + sampling_season + gender + smoked_ever, data = meta_op_cp_gf) %>% 
  broom::tidy(., conf.int = T)
lm_shannon_op_cp_gf$livestock <- "Occupational exposure (GF vs CP)"
lm_shannon_op_cp_gf$niche <- "OP"


# Linear regression for observed diversity
lm_observed_np_cp_gf <- lm(Observed ~ population + age + sampling_season + gender + smoked_ever, data = meta_np_cp_gf) %>% 
  broom::tidy(., conf.int = T)
lm_observed_np_cp_gf$livestock <- "Occupational exposure (GF vs CP)"
lm_observed_np_cp_gf$niche <- "NP"

lm_observed_op_cp_gf <- lm(Observed ~ population + age + sampling_season + gender + smoked_ever, data = meta_op_cp_gf) %>% 
  broom::tidy(., conf.int = T)
lm_observed_op_cp_gf$livestock <- "Occupational exposure (GF vs CP)"
lm_observed_op_cp_gf$niche <- "OP"


# Combine all the tidy data
combined_shannon_cp_gf <- bind_rows(lm_shannon_np_cp_gf, lm_shannon_op_cp_gf)
# Filter to include only livestock exposure terms
terms_shannon_cp_gf <- combined_shannon_cp_gf %>%
  filter(term %in% c("populationGF"))
# Add adjusted p-value labels
terms_shannon_cp_gf <- terms_shannon_cp_gf %>% 
  mutate(
    label = case_when(
      p.value < 0.1 ~ "<0.1",  # Assign "<0.1" if p.value is less than 0.1
      TRUE ~ "NS"              # Otherwise, assign "NS"
    )
  )


write.csv(terms_shannon_cp_gf,"../Output/Alpha_diversity/shannon_cp_gf.csv")


# Combine all the tidy data
combined_observed_cp_gf <- bind_rows(lm_observed_np_cp_gf, lm_observed_op_cp_gf)

# Filter to include only livestock exposure terms
terms_observed_cp_gf <- combined_observed_cp_gf %>%
  filter(term %in% c("populationGF"))
# Add adjusted p-value labels
terms_observed_cp_gf <- terms_observed_cp_gf %>% 
  mutate(
    label = case_when(
      p.value < 0.1 ~ "<0.1",  # Assign "<0.1" if p.value is less than 0.1
      TRUE ~ "NS"              # Otherwise, assign "NS"
    )
  )



write.csv(terms_observed_cp_gf,"../Output/Alpha_diversity/observedrichness_cp_gf.csv", )


# Combine original metadata dataframes together 
metadata_combined_np_op <- bind_rows(meta_np_cp_gf, meta_op_cp_gf)
metadata_combined_np_op


# Set up custom colors for CP and GF with darker shades for edges, median line, and outliers
color_cp <- "#e3be6b"
color_gf <- "#76b5a3"
darken_color <- function(color, factor = 0.6) {
  color_rgb <- col2rgb(color) / 255  # Convert to RGB values (0 to 1 scale)
  darkened_rgb <- color_rgb * factor  # Darken by multiplying by factor
  darkened_color <- rgb(darkened_rgb[1], darkened_rgb[2], darkened_rgb[3])  # Convert back to color
  return(darkened_color)
}
color_cp_dark <- darken_color(color_cp)
color_gf_dark <- darken_color(color_gf)

# Boxplot for Shannon diversity with significance
shannon_boxplot <- ggplot(metadata_combined_np_op, aes(x = population, y = Shannon, fill = population, color = population)) +
  geom_boxplot(
    outlier.colour = "black",  # Color for outliers
    outlier.size = 3,          # Size of outliers
    outlier.shape = 16         # Shape of outliers
  ) +
  facet_wrap(~ niche) +
  labs(x = "Population", y = "Shannon Diversity Index") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none", 
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),
    axis.text = element_text(size = 12),
    plot.title = element_text(size = 18)
  ) +
  scale_fill_manual(values = c("CP" = color_cp, "GF" = color_gf)) +  # Custom colors for CP and GF
  scale_color_manual(values = c("CP" = color_cp_dark, "GF" = color_gf_dark)) +  # Darker edges and median for CP and GF
  geom_signif(
    comparisons = list(c("CP", "GF")),  # Specify the comparison
    annotations = terms_shannon_cp_gf$label[1],  # Match the label for Shannon significance
    map_signif_level = FALSE,  # Turn off automatic labeling
    tip_length = 0.03,
    size = 1,
    textsize = 3,
    color = "black" 
  )

shannon_boxplot

# Boxplot for Observed diversity with significance
observed_boxplot <- ggplot(metadata_combined_np_op, aes(x = population, y = Observed, fill = population, color = population)) +
  geom_boxplot(
    outlier.colour = "black",  # Color for outliers
    outlier.size = 3,          # Size of outliers
    outlier.shape = 16         # Shape of outliers
  ) +
  facet_wrap(~ niche) +
  labs(x = "Population", y = "Observed Diversity Index") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none", 
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),
    axis.text = element_text(size = 12),
    plot.title = element_text(size = 18)
  ) +
  scale_fill_manual(values = c("CP" = color_cp, "GF" = color_gf)) +  # Custom colors for CP and GF
  scale_color_manual(values = c("CP" = color_cp_dark, "GF" = color_gf_dark)) +  # Darker edges and median for CP and GF
  geom_signif(
    comparisons = list(c("CP", "GF")),  # Specify the comparison
    annotations = terms_observed_cp_gf$label[1],  # Match the label for Observed significance
    map_signif_level = FALSE,  # Turn off automatic labeling
    tip_length = 0.03,
    size = 1,
    textsize = 3,
    color = "black"  # Set the p-value label text color to black
  )
observed_boxplot
# Save the boxplots
ggsave("../Output/Alpha_diversity/Shannon_GFvsCP.svg", shannon_boxplot, width = 4, height = 4)
ggsave("../Output/Alpha_diversity/Observed_GFvsCP.svg", observed_boxplot, width = 4, height = 4)




# Violin plots
# Violin + Boxplot for Shannon Diversity with significance
shannon_violin_boxplot <- ggplot(metadata_combined_np_op, aes(x = population, y = Shannon, fill = population, color = population)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8, color = "black") +  # Boxplot inside violin
  facet_wrap(~ niche) +
  labs(x = "Population", y = "Shannon Diversity") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none", 
    panel.border = element_rect(color = "black", fill = NA),
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),
    axis.title = element_text(size = 14),
    axis.text = element_text(size = 12),
    plot.title = element_text(size = 18)
  ) +
  scale_fill_manual(values = c("CP" = color_cp, "GF" = color_gf)) +  # Violin fill color
  scale_color_manual(values = c("CP" = color_cp_dark, "GF" = color_gf_dark)) +  # Boxplot edges
  geom_signif(
    comparisons = list(c("CP", "GF")),  
    annotations = terms_shannon_cp_gf$label[1],  
    map_signif_level = FALSE,  
    tip_length = 0.03,
    size = 1,
    textsize = 3,
    color = "black"  
  )

# Violin + Boxplot for Observed Diversity with significance
observed_violin_boxplot <- ggplot(metadata_combined_np_op, aes(x = population, y = Observed, fill = population, color = population)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8, color = "black") +  # Boxplot inside violin
  facet_wrap(~ niche) +
  labs(x = "Population", y = "Observed richness") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none", 
    panel.border = element_rect(color = "black", fill = NA),
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),
    axis.title = element_text(size = 14),
    axis.text = element_text(size = 12),
    plot.title = element_text(size = 18)
  ) +
  scale_fill_manual(values = c("CP" = color_cp, "GF" = color_gf)) +  # Violin fill color
  scale_color_manual(values = c("CP" = color_cp_dark, "GF" = color_gf_dark)) +  # Boxplot edges
  geom_signif(
    comparisons = list(c("CP", "GF")),  
    annotations = terms_observed_cp_gf$label[1],  
    map_signif_level = FALSE,  
    tip_length = 0.03,
    size = 1,
    textsize = 3,
    color = "black"  
  )

# Save the violin + boxplots
ggsave("../Output/Alpha_diversity/Shannon_GFvsCP_violin_boxplot.svg", shannon_violin_boxplot, width = 4, height = 4)
ggsave("../Output/Alpha_diversity/Observed_GFvsCP_violin_boxplot.svg", observed_violin_boxplot, width = 4, height = 4)



```

## Linear models on 16S qPCR

```{r}
# Linear regression for shannon 
lm_qPCR_np_cp_gf <- lm(qPCR ~ population + age + sampling_season + gender + smoked_ever, data = meta_np_cp_gf) %>% 
  broom::tidy(., conf.int = T)
lm_qPCR_np_cp_gf$livestock <- "Occupational exposure (GF vs CP)"
lm_qPCR_np_cp_gf$niche <- "NP"

lm_qPCR_op_cp_gf <- lm(qPCR ~ population + age + sampling_season + gender + smoked_ever, data = meta_op_cp_gf) %>% 
  broom::tidy(., conf.int = T)
lm_qPCR_op_cp_gf$livestock <- "Occupational exposure (GF vs CP)"
lm_qPCR_op_cp_gf$niche <- "OP"


# Combine all the tidy data
combined_qPCR_cp_gf <- bind_rows(lm_qPCR_np_cp_gf, lm_qPCR_op_cp_gf)
# Filter to include only livestock exposure terms
terms_qPCR_cp_gf <- combined_qPCR_cp_gf %>%
  filter(term %in% c("populationGF"))

# Add adjusted p-value labels
terms_qPCR_cp_gf <- terms_qPCR_cp_gf %>% 
  mutate(
    label = case_when(
      p.value < 0.1 ~ "<0.1",  # Assign "<0.1" if p.value is less than 0.1
      TRUE ~ "NS"              # Otherwise, assign "NS"
    )
  )

write.csv(terms_qPCR_cp_gf,"../Output/Alpha_diversity/terms_qPCR_cp_gf.csv", )

# Combine original metadata dataframes together 
metadata_combined_np_op <- bind_rows(meta_np_cp_gf, meta_op_cp_gf)
metadata_combined_np_op


# Set up custom colors for CP and GF with darker shades for edges, median line, and outliers
color_cp <- "#e3be6b"
color_gf <- "#76b5a3"
darken_color <- function(color, factor = 0.6) {
  color_rgb <- col2rgb(color) / 255  # Convert to RGB values (0 to 1 scale)
  darkened_rgb <- color_rgb * factor  # Darken by multiplying by factor
  darkened_color <- rgb(darkened_rgb[1], darkened_rgb[2], darkened_rgb[3])  # Convert back to color
  return(darkened_color)
}
color_cp_dark <- darken_color(color_cp)
color_gf_dark <- darken_color(color_gf)

# Apply log transformation (ensure you handle zeros properly)
metadata_combined_np_op$log_qPCR <- log1p(metadata_combined_np_op$qPCR)


metadata_combined_np_op$qPCR_log
# Violin + Boxplot for Observed Diversity with significance
qPCR_violin_boxplot <- ggplot(metadata_combined_np_op, aes(x = population, y = log_qPCR, fill = population, color = population)) +
  geom_violin(trim = TRUE, alpha = 0.6, color = NA) +  # Violin plot
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.8, color = "black") +  # Boxplot inside violin
  facet_wrap(~ niche) +
  labs(x = "Population", y = "16S qPCR") +
  theme_minimal(base_size = 16) +
  theme(
    legend.position = "none", 
    panel.border = element_rect(color = "black", fill = NA),
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),
    axis.title = element_text(size = 14),
    axis.text = element_text(size = 12),
    plot.title = element_text(size = 18)
  ) +
  scale_fill_manual(values = c("CP" = color_cp, "GF" = color_gf)) +  # Violin fill color
  scale_color_manual(values = c("CP" = color_cp_dark, "GF" = color_gf_dark)) +  # Boxplot edges
  geom_signif(
    comparisons = list(c("CP", "GF")),  # Specify the comparison
    annotations = terms_qPCR_cp_gf$label[1],
    map_signif_level = FALSE,  # Turn off automatic labeling
    tip_length = 0.03,
    size = 1,
    textsize = 3,
    color = "black" 
  )

# Save
ggsave("../Output/Alpha_diversity/qPCR_GFvsCP.svg", qPCR_violin_boxplot, width = 4, height = 4)
```

## Figure for supplement
```{r}
# Plot for supplement
# Supplementary figure (observed richness and qPCR)
population_obs_qPCR_plot <- plot_grid(observed_violin_boxplot, qPCR_violin_boxplot, nrow = 1, align = "hv")
population_obs_qPCR_plot



```

# Residential livestock exposure 
Run linear models associating the diversity indices (Shannon and observed richness) and qPCR with the residential livestock exposure variables as input
## Non-standardised livestock variables
```{r}
# We use linear models to test the effect of various metadata variables on the Shannon index, observed species and bacterial density. This was done for each niche separately. Models were corrected for age, season of sampling, sex and smoking status of the participant. We visualize the results as boxplots including the p-values from the tests.

#### Shannon ####
# NP niche
lm_animals3km_np <- lm(Shannon ~  KRD_nCowsWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_animals3km_np$niche <- "NP"
# Loop through the coefficients and add the 'livestock' label based on the variable name
lm_animals3km_np$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum", lm_animals3km_np$term), "Cattle", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum", lm_animals3km_np$term), "Goats", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum", lm_animals3km_np$term), "Pigs", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum", lm_animals3km_np$term), "Poultry", NA))))



# modelled microbial agents
lm_endoexpo_np <- lm(Shannon ~  EU_ANNUALAVG_PM10_2020 + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_np$livestock <- "Dispersion-modelled endotoxin"
lm_endoexpo_np$niche <- "NP"


lm_ecoliexpo_np <- lm(Shannon ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_np$livestock <- "RF-modelled E. coli"
lm_ecoliexpo_np$niche <- "NP"


lm_staphexpo_np <- lm(Shannon ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_np$livestock <- "RF-modelled Staph"
lm_staphexpo_np$niche <- "NP"


lm_tetwexpo_np <- lm(Shannon ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_np$livestock <- "RF-modelled tetW"
lm_tetwexpo_np$niche <- "NP"

lm_mecaexpo_np <- lm(Shannon ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_np$livestock <- "RF-modelled mecA"
lm_mecaexpo_np$niche <- "NP"


# OP niche
lm_animals3km_op <- lm(Shannon ~  KRD_nCowsWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_op_cp) %>% 
  broom::tidy(., conf.int = T)
lm_animals3km_op$niche <- "OP"

lm_animals3km_op$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum", lm_animals3km_op$term), "Cattle", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum", lm_animals3km_op$term), "Goats", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum", lm_animals3km_op$term), "Pigs", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum", lm_animals3km_op$term), "Poultry", NA))))


# modelled microbial agents
lm_endoexpo_op <- lm(Shannon ~  EU_ANNUALAVG_PM10_2020 + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_op$livestock <- "Dispersion-modelled endotoxin"
lm_endoexpo_op$niche <- "OP"

lm_ecoliexpo_op <- lm(Shannon ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_op$livestock <- "RF-modelled E. coli"
lm_ecoliexpo_op$niche <- "OP"

lm_staphexpo_op <- lm(Shannon ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_op$livestock <- "RF-modelled Staph"
lm_staphexpo_op$niche <- "OP"

lm_tetwexpo_op <- lm(Shannon ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_op$livestock <- "RF-modelled tetW"
lm_tetwexpo_op$niche <- "OP"

lm_mecaexpo_op <- lm(Shannon ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_op$livestock <- "RF-modelled mecA"
lm_mecaexpo_op$niche <- "OP"


# Combine all the tidy data
combined_data_shannon <- bind_rows(lm_animals3km_np, lm_endoexpo_np, lm_ecoliexpo_np, lm_staphexpo_np, lm_tetwexpo_np, lm_mecaexpo_np, lm_animals3km_op, lm_endoexpo_op, lm_ecoliexpo_op, lm_staphexpo_op, lm_tetwexpo_op, lm_mecaexpo_op)

# Filter to include only livestock exposure terms
livestock_terms_shannon <- combined_data_shannon %>%
  filter(term %in% c("KRD_nCowsWghtDist.3000m.sum", 
                     "KRD_nGoatsWghtDist.3000m.sum", 
                     "KRD_nPigsWghtDist.3000m.sum", 
                     "KRD_nPoultryWghtDist.3000m.sum",
                     "EU_ANNUALAVG_PM10_2020",
                     "ecoli_RF_preds",
                     "staph_RF_preds",
                     "meca_RF_preds",
                     "tetw_RF_preds"))

# Apply multiple testing correction (Benjamini-Hochberg) for NP
livestock_terms_np_shannon <- livestock_terms_shannon %>%
  filter(niche == "NP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Apply multiple testing correction (Benjamini-Hochberg) for OP
livestock_terms_op_shannon <- livestock_terms_shannon %>%
  filter(niche == "OP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))
    
# Combine the results for NP and OP niches
livestock_terms_corrected_shannon <- bind_rows(livestock_terms_np_shannon, livestock_terms_op_shannon)

# Add adjusted p-value labels
livestock_terms_corrected_shannon <- livestock_terms_corrected_shannon %>%
  mutate(
    label = case_when(
      p.adjusted < 0.1 ~ "*",  # If p.adjusted is less than 0.1, assign *
      TRUE ~ NA_character_     # Otherwise, set label to NA
    )
  )

write.csv(livestock_terms_corrected_shannon,"../Output/Alpha_diversity/Residential_livestock_expo_shannondiversity_linear_regression.csv")

# Define the desired order of the variables and their new names
reordered_variables <- c(
  "Cattle" = "Cattle",
  "Poultry" = "Chickens",
  "Goats" = "Goats",
  "Pigs" = "Pigs",
  "RF-modelled mecA" = "mecA exposure",
  "RF-modelled tetW" = "tetW exposure",
  "RF-modelled Staph" = "Staph spp. exposure",
  "RF-modelled E. coli" = "E. coli exposure",
  "Dispersion-modelled endotoxin" = "Endotoxin exposure"

)

# Reorder and rename the levels of 'livestock' variable
livestock_terms_corrected_shannon$livestock <- factor(
  livestock_terms_corrected_shannon$livestock,
  levels = names(reordered_variables),
  labels = reordered_variables
)


# Shannon
livestock_expo_shannon <- ggplot(livestock_terms_corrected_shannon, aes(x = livestock, y = estimate, color = niche)) +
  geom_point(position = position_dodge(width = 0.5), size = 3) +  # Plot points
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.2, 
                position = position_dodge(width = 0.5)) +  # Error bars for confidence intervals
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +  # Add dotted line at 0
  labs(x = "Livestock exposure", 
       y = "Shannon Diversity Regression Coefficient",
       color = "Niche") +
  theme_minimal(base_size = 16) +  # Increase base font size
  theme(
    legend.position = "top",  # Move legend to the top
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),  # Increase size of axis titles
    axis.text = element_text(size = 12),   # Increase size of axis text
    plot.title = element_text(size = 18)   # Increase size of the plot title
  ) +
  scale_color_manual(values = c("NP" = "#dd968e", "OP" = "#547484")) +  # Custom colors for NP and OP
  facet_wrap(~ niche, scales = "free_x") +  # Facet by niche (NP and OP)
  coord_flip()   # Flip coordinates for better readability
livestock_expo_shannon


# Save the plot
theme(
  plot.margin = margin(1, 1, 2, 1, "cm")  # More space at the bottom (2 cm)
)

ggsave("../Output/Alpha_diversity/livestock_expo_shannon.svg", livestock_expo_shannon, width = 6, height = 5)


#### Observed ####
# We use linear models to test the effect of various metadata variables on the Observed index, Observed species and bacterial density. This was done for each niche separately. Models were corrected for age, season of sampling, sex and smoking status of the participant. We visualize the results as boxplots including the p-values from the tests.
# NP niche
lm_animals3km_np_observed <- lm(Observed ~  KRD_nCowsWghtDist.3000m.sum +KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_animals3km_np_observed$niche <- "NP"
lm_animals3km_np_observed$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum", lm_animals3km_np_observed$term), "Cattle", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum", lm_animals3km_np_observed$term), "Goats", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum", lm_animals3km_np_observed$term), "Pigs", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum", lm_animals3km_np_observed$term), "Poultry", NA))))


# RF- modelled microbial agents
lm_endoexpo_np_observed <- lm(Observed ~  EU_ANNUALAVG_PM10_2020 + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_np_observed$livestock <- "Dispersion-modelled endotoxin"
lm_endoexpo_np_observed$niche <- "NP"


lm_ecoliexpo_np_observed <- lm(Observed ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_np_observed$livestock <- "RF-modelled E. coli"
lm_ecoliexpo_np_observed$niche <- "NP"


lm_staphexpo_np_observed <- lm(Observed ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_np_observed$livestock <- "RF-modelled Staph"
lm_staphexpo_np_observed$niche <- "NP"


lm_tetwexpo_np_observed <- lm(Observed ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_np_observed$livestock <- "RF-modelled tetW"
lm_tetwexpo_np_observed$niche <- "NP"

lm_mecaexpo_np_observed <- lm(Observed ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_np_observed$livestock <- "RF-modelled mecA"
lm_mecaexpo_np_observed$niche <- "NP"


# OP niche
lm_animals3km_op_observed <- lm(Observed ~  KRD_nCowsWghtDist.3000m.sum +KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_animals3km_op_observed$niche <- "OP"
lm_animals3km_op_observed$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum", lm_animals3km_op_observed$term), "Cattle", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum", lm_animals3km_op_observed$term), "Goats", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum", lm_animals3km_op_observed$term), "Pigs", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum", lm_animals3km_op_observed$term), "Poultry", NA))))


# RF modelled microbial agents
lm_endoexpo_op_observed <- lm(Observed ~  EU_ANNUALAVG_PM10_2020 + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_op_observed$livestock <- "Dispersion-modelled endotoxin"
lm_endoexpo_op_observed$niche <- "OP"


lm_ecoliexpo_op_observed <- lm(Shannon ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_op_observed$livestock <- "RF-modelled E. coli"
lm_ecoliexpo_op_observed$niche <- "OP"

lm_staphexpo_op_observed <- lm(Shannon ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_op_observed$livestock <- "RF-modelled Staph"
lm_staphexpo_op_observed$niche <- "OP"

lm_tetwexpo_op_observed <- lm(Shannon ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_op_observed$livestock <- "RF-modelled tetW"
lm_tetwexpo_op_observed$niche <- "OP"

lm_mecaexpo_op_observed <- lm(Shannon ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_op_observed$livestock <- "RF-modelled mecA"
lm_mecaexpo_op_observed$niche <- "OP"

# Combine all the tidy data
combined_data_observed <- bind_rows(lm_animals3km_np_observed, lm_ecoliexpo_np_observed, lm_endoexpo_np_observed, lm_staphexpo_np_observed, lm_tetwexpo_np_observed, lm_mecaexpo_np_observed, lm_animals3km_op_observed, lm_endoexpo_op_observed, lm_ecoliexpo_op_observed, lm_staphexpo_op_observed, lm_tetwexpo_op_observed, lm_mecaexpo_op_observed)

# Filter to include only livestock exposure terms (e.g., KRD variables)
livestock_terms_observed <- combined_data_observed %>%
  filter(term %in% c("KRD_nCowsWghtDist.3000m.sum", 
                     "KRD_nGoatsWghtDist.3000m.sum", 
                     "KRD_nPigsWghtDist.3000m.sum", 
                     "KRD_nPoultryWghtDist.3000m.sum",
                     "EU_ANNUALAVG_PM10_2020",
                     "ecoli_RF_preds",
                     "staph_RF_preds",
                     "tetw_RF_preds",
                     "meca_RF_preds"))


# Apply multiple testing correction (Benjamini-Hochberg) for NP
livestock_terms_np_observed <- livestock_terms_observed %>%
  filter(niche == "NP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Apply multiple testing correction (Benjamini-Hochberg) for OP
livestock_terms_op_observed <- livestock_terms_observed %>%
  filter(niche == "OP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Combine the results for NP and OP niches
livestock_terms_corrected_observed <- bind_rows(livestock_terms_np_observed, livestock_terms_op_observed)

# Add adjusted p-value labels
livestock_terms_corrected_observed <- livestock_terms_corrected_observed %>%
  mutate(
    label = case_when(
      p.adjusted < 0.1 ~ "*",  # If p.adjusted is less than 0.1, assign *
      TRUE ~ NA_character_     # Otherwise, set label to NA
    )
  )


write.csv(livestock_terms_corrected_observed,"../Output/Alpha_diversity/Residential_livestock_expo_observedrichness_linear_regression.csv")

# Define the desired order of the variables and their new names
reordered_variables <- c(
  "Cattle" = "Cattle",
  "Poultry" = "Chickens",
  "Goats" = "Goats",
  "Pigs" = "Pigs",
  "RF-modelled mecA" = "mecA exposure",
  "RF-modelled tetW" = "tetW exposure",
  "RF-modelled Staph" = "Staph spp. exposure",
  "RF-modelled E. coli" = "E. coli exposure",
  "Dispersion-modelled endotoxin"= "Endotoxin exposure"
)

# Reorder and rename the levels of 'livestock' variable
livestock_terms_corrected_observed$livestock <- factor(
  livestock_terms_corrected_observed$livestock,
  levels = names(reordered_variables),
  labels = reordered_variables
)

# observed plot
livestock_expo_observed <- ggplot(livestock_terms_corrected_observed, aes(x = livestock, y = estimate, color = niche)) +
  geom_point(position = position_dodge(width = 0.5), size = 3) +  # Plot points
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.2, 
                position = position_dodge(width = 0.5)) +  # Error bars for confidence intervals
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +  # Add dotted line at 0
  labs(x = "Livestock exposure", 
       y = "Observed Diversity Regression Coefficient",
       color = "Niche") +
  theme_minimal(base_size = 16) +  # Increase base font size
  theme(
    legend.position = "top",  # Move legend to the top
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),  # Increase size of axis titles
    axis.text = element_text(size = 12),   # Increase size of axis text
    plot.title = element_text(size = 18)   # Increase size of the plot title
  ) +
  scale_color_manual(values = c("NP" = "#dd968e", "OP" = "#547484")) +  # Custom colors for NP and OP
  facet_wrap(~ niche, scales = "free_x") +  # Facet by niche (NP and OP)
  coord_flip()   # Flip coordinates for better readability
livestock_expo_observed

# Save 
ggsave("../Output/Alpha_diversity/livestock_expo_observed.svg", livestock_expo_observed,width = 6, height=5 )



#### qPCR ####
# NP niche
lm_cattle2km_np_qPCR <- lm(qPCR ~  KRD_nCowsWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_cattle2km_np_qPCR$livestock <- "Cattle"
lm_cattle2km_np_qPCR$niche <- "NP"

lm_goats2km_np_qPCR <- lm(qPCR ~  KRD_nGoatsWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_goats2km_np_qPCR$livestock <- "Goats"
lm_goats2km_np_qPCR$niche <- "NP"

lm_pigs2km_np_qPCR <- lm(qPCR ~  KRD_nPigsWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_pigs2km_np_qPCR$livestock <- "Pigs"
lm_pigs2km_np_qPCR$niche <- "NP"

lm_poultry2km_np_qPCR <- lm(qPCR ~  KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_poultry2km_np_qPCR$livestock <- "Poultry"
lm_poultry2km_np_qPCR$niche <- "NP"

# RF- modelled microbial agents
lm_endoexpo_np_qPCR <- lm(qPCR ~  EU_ANNUALAVG_PM10_2020 + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_np_qPCR$livestock <- "Dispersion-modelled endotoxin"
lm_endoexpo_np_qPCR$niche <- "NP"


lm_ecoliexpo_np_qPCR <- lm(qPCR ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_np_qPCR$livestock <- "RF-modelled E. coli"
lm_ecoliexpo_np_qPCR$niche <- "NP"


lm_staphexpo_np_qPCR <- lm(qPCR ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_np_qPCR$livestock <- "RF-modelled Staph"
lm_staphexpo_np_qPCR$niche <- "NP"


lm_tetwexpo_np_qPCR <- lm(qPCR ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_np_qPCR$livestock <- "RF-modelled tetW"
lm_tetwexpo_np_qPCR$niche <- "NP"

lm_mecaexpo_np_qPCR <- lm(qPCR ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_np_qPCR$livestock <- "RF-modelled mecA"
lm_mecaexpo_np_qPCR$niche <- "NP"


# OP niche
lm_cattle2km_op_qPCR <- lm(qPCR ~  KRD_nCowsWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_op_cp) %>% 
  broom::tidy(., conf.int = T)
lm_cattle2km_op_qPCR$livestock <- "Cattle"
lm_cattle2km_op_qPCR$niche <- "OP"

lm_goats2km_op_qPCR <- lm(qPCR ~  KRD_nGoatsWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_goats2km_op_qPCR$livestock <- "Goats"
lm_goats2km_op_qPCR$niche <- "OP"

lm_pigs2km_op_qPCR <- lm(qPCR ~  KRD_nPigsWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_pigs2km_op_qPCR$livestock <- "Pigs"
lm_pigs2km_op_qPCR$niche <- "OP"

lm_poultry2km_op_qPCR <- lm(qPCR ~  KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>%
  broom::tidy(., conf.int = T)
lm_poultry2km_op_qPCR$livestock <- "Poultry"
lm_poultry2km_op_qPCR$niche <- "OP"

# RF modelled microbial agents
lm_endoexpo_op_qPCR <- lm(qPCR ~  EU_ANNUALAVG_PM10_2020 + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_op_qPCR$livestock <- "Dispersion-modelled endotoxin"
lm_endoexpo_op_qPCR$niche <- "OP"

lm_ecoliexpo_op_qPCR <- lm(qPCR ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_op_qPCR$livestock <- "RF-modelled E. coli"
lm_ecoliexpo_op_qPCR$niche <- "OP"

lm_staphexpo_op_qPCR <- lm(qPCR ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_op_qPCR$livestock <- "RF-modelled Staph"
lm_staphexpo_op_qPCR$niche <- "OP"

lm_tetwexpo_op_qPCR <- lm(qPCR ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_op_qPCR$livestock <- "RF-modelled tetW"
lm_tetwexpo_op_qPCR$niche <- "OP"

lm_mecaexpo_op_qPCR <- lm(qPCR ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_op_qPCR$livestock <- "RF-modelled mecA"
lm_mecaexpo_op_qPCR$niche <- "OP"


# Combine all the tidy data
combined_data_qPCR <- bind_rows(lm_cattle2km_np_qPCR, lm_goats2km_np_qPCR, lm_pigs2km_np_qPCR, lm_poultry2km_np_qPCR, lm_endoexpo_np_qPCR, lm_ecoliexpo_np_qPCR, lm_staphexpo_np_qPCR, lm_tetwexpo_np_qPCR, lm_mecaexpo_np_qPCR, lm_cattle2km_op_qPCR, lm_goats2km_op_qPCR, lm_pigs2km_op_qPCR, lm_poultry2km_op_qPCR,lm_endoexpo_op_qPCR, lm_ecoliexpo_op_qPCR, lm_staphexpo_op_qPCR, lm_tetwexpo_op_qPCR, lm_mecaexpo_op_qPCR)

# Filter to include only livestock exposure terms (e.g., KRD variables)
livestock_terms_qPCR <- combined_data_qPCR %>%
  filter(term %in% c("KRD_nCowsWghtDist.3000m.sum", 
                     "KRD_nGoatsWghtDist.3000m.sum", 
                     "KRD_nPigsWghtDist.3000m.sum", 
                     "KRD_nPoultryWghtDist.3000m.sum",
                     "EU_ANNUALAVG_PM10_2020",
                     "ecoli_RF_preds",
                     "staph_RF_preds",
                     "tetw_RF_preds",
                     "meca_RF_preds"))


# Apply multiple testing correction (Benjamini-Hochberg) for NP
livestock_terms_np_qPCR <- livestock_terms_qPCR %>%
  filter(niche == "NP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Apply multiple testing correction (Benjamini-Hochberg) for OP
livestock_terms_op_qPCR <- livestock_terms_qPCR %>%
  filter(niche == "OP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Combine the results for NP and OP niches
livestock_terms_corrected_qPCR <- bind_rows(livestock_terms_np_qPCR, livestock_terms_op_qPCR)

# Add adjusted p-value labels
livestock_terms_corrected_qPCR <- livestock_terms_corrected_qPCR %>%
  mutate(
    label = case_when(
      p.adjusted < 0.1 ~ "*",  # If p.adjusted is less than 0.1, assign *
      TRUE ~ NA_character_     # Otherwise, set label to NA
    )
  )


write.csv(livestock_terms_corrected_qPCR,"../Output/Alpha_diversity/Residential_livestock_expo_qPCR_linear_regression.csv")

# Define the desired order of the variables and their new names
reordered_variables <- c(
  "Cattle" = "Cattle",
  "Poultry" = "Chickens",
  "Goats" = "Goats",
  "Pigs" = "Pigs",
  "RF-modelled mecA" = "mecA exposure",
  "RF-modelled tetW" = "tetW exposure",
  "RF-modelled Staph" = "Staph spp. exposure",
  "RF-modelled E. coli" = "E. coli exposure",
  "Dispersion-modelled endotoxin" = "Endotoxin exposure")


# Reorder and rename the levels of 'livestock' variable
livestock_terms_corrected_qPCR$livestock <- factor(
  livestock_terms_corrected_qPCR$livestock,
  levels = names(reordered_variables),
  labels = reordered_variables
)

# qPCR plot
livestock_expo_qPCR <- ggplot(livestock_terms_corrected_qPCR, aes(x = livestock, y = estimate, color = niche)) +
  geom_point(position = position_dodge(width = 0.5), size = 3) +  # Plot points
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.2, 
                position = position_dodge(width = 0.5)) +  # Error bars for confidence intervals
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +  # Add dotted line at 0
  geom_text(aes(label = label), 
            position = position_dodge(width = 0.5), 
            vjust = -0.5, size = 4, color = "black") +  # Add labels above the points
  labs(x = "Livestock exposure", 
       y = "16S qPCR",
       color = "Niche") +
  theme_minimal(base_size = 16) +  # Increase base font size
  theme(
    legend.position = "top",  # Move legend to the top
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),  # Increase size of axis titles
    axis.text = element_text(size = 12),   # Increase size of axis text
    plot.title = element_text(size = 18)   # Increase size of the plot title
  ) +
  scale_color_manual(values = c("NP" = "#dd968e", "OP" = "#547484")) +  # Custom colors for NP and OP
  facet_wrap(~ niche, scales = "free_x") +  # Facet by niche (NP and OP)
  coord_flip()   # Flip coordinates for better readability
livestock_expo_qPCR

# Save 
ggsave("../Output/Alpha_diversity/livestock_expo_qPCR.svg", livestock_expo_qPCR,width = 6, height=5)
```
### Model assumption checking
```{r}
# Shannon models
# animals 3km
lm_animals3km_np <- lm(Shannon ~ KRD_nCowsWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_animals3km_np)
residuals_values <- residuals(lm_animals3km_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_animals3km_np$residuals)
qqline(lm_animals3km_np$residuals, col = "red")
hist(lm_animals3km_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_animals3km_np$fitted.values, abs(lm_animals3km_np$residuals))
abline(h = 0, col = "red")
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_animals3km_np) # no autocorrelation
# multicollinearity
vif(lm_animals3km_np)
# Outliers
plot(lm_animals3km_np, which = 4)  # Cook's distance plot


# ecoli 3km
lm_ecoliexpo_np <- lm(Shannon ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_ecoliexpo_np)
residuals_values <- residuals(lm_ecoliexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_ecoliexpo_np$residuals)
qqline(lm_ecoliexpo_np$residuals, col = "red")
hist(lm_ecoliexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_ecoliexpo_np$fitted.values, abs(lm_ecoliexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_ecoliexpo_np) # no autocorrelation
# multicollinearity
vif(lm_ecoliexpo_np)
# Outliers
plot(lm_ecoliexpo_np, which = 4)  # Cook's distance plot



# Staph model
lm_staphexpo_np <- lm(Shannon ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_staphexpo_np)
residuals_values <- residuals(lm_staphexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_staphexpo_np$residuals)
qqline(lm_staphexpo_np$residuals, col = "red")
hist(lm_staphexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_staphexpo_np$fitted.values, abs(lm_staphexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_staphexpo_np) # no autocorrelation
# multicollinearity
vif(lm_staphexpo_np)
# Outliers
plot(lm_staphexpo_np, which = 4)  # Cook's distance plot



# tetw model
lm_tetwexpo_np <- lm(Shannon ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_tetwexpo_np)
residuals_values <- residuals(lm_tetwexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_tetwexpo_np$residuals)
qqline(lm_tetwexpo_np$residuals, col = "red")
hist(lm_tetwexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_tetwexpo_np$fitted.values, abs(lm_tetwexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_tetwexpo_np) # no autocorrelation
# multicollinearity
vif(lm_tetwexpo_np)
# Outliers
plot(lm_tetwexpo_np, which = 4)  # Cook's distance plot


# meca model
lm_mecaexpo_np <- lm(Shannon ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_mecaexpo_np)
residuals_values <- residuals(lm_mecaexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_mecaexpo_np$residuals)
qqline(lm_mecaexpo_np$residuals, col = "red")
hist(lm_mecaexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_mecaexpo_np$fitted.values, abs(lm_mecaexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_mecaexpo_np) # no autocorrelation
# multicollinearity
vif(lm_mecaexpo_np)
# Outliers
plot(lm_mecaexpo_np, which = 4)  # Cook's distance plot






# Observed diversity models
# animals 3km
lm_animals3km_np <- lm(Observed ~ KRD_nCowsWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nPoultryWghtDist.3000m.sum + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_animals3km_np)
residuals_values <- residuals(lm_animals3km_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_animals3km_np$residuals)
qqline(lm_animals3km_np$residuals, col = "red")
hist(lm_animals3km_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_animals3km_np$fitted.values, abs(lm_animals3km_np$residuals))
abline(h = 0, col = "red")
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_animals3km_np) # no autocorrelation
# multicollinearity
vif(lm_animals3km_np)
# Outliers
plot(lm_animals3km_np, which = 4)  # Cook's distance plot


# ecoli 3km
lm_ecoliexpo_np <- lm(Observed ~  ecoli_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_ecoliexpo_np)
residuals_values <- residuals(lm_ecoliexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_ecoliexpo_np$residuals)
qqline(lm_ecoliexpo_np$residuals, col = "red")
hist(lm_ecoliexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_ecoliexpo_np$fitted.values, abs(lm_ecoliexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_ecoliexpo_np) # no autocorrelation
# multicollinearity
vif(lm_ecoliexpo_np)
# Outliers
plot(lm_ecoliexpo_np, which = 4)  # Cook's distance plot



# Staph model
lm_staphexpo_np <- lm(Observed ~  staph_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_staphexpo_np)
residuals_values <- residuals(lm_staphexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_staphexpo_np$residuals)
qqline(lm_staphexpo_np$residuals, col = "red")
hist(lm_staphexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_staphexpo_np$fitted.values, abs(lm_staphexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_staphexpo_np) # no autocorrelation
# multicollinearity
vif(lm_staphexpo_np)
# Outliers
plot(lm_staphexpo_np, which = 4)  # Cook's distance plot



# tetw model
lm_tetwexpo_np <- lm(Observed ~  tetw_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_tetwexpo_np)
residuals_values <- residuals(lm_tetwexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_tetwexpo_np$residuals)
qqline(lm_tetwexpo_np$residuals, col = "red")
hist(lm_tetwexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_tetwexpo_np$fitted.values, abs(lm_tetwexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_tetwexpo_np) # no autocorrelation
# multicollinearity
vif(lm_tetwexpo_np)
# Outliers
plot(lm_tetwexpo_np, which = 4)  # Cook's distance plot


# meca model
lm_mecaexpo_np <- lm(Observed ~  meca_RF_preds + age + sampling_season + gender + smoked_ever, data = meta_np_cp)
fitted_values <- fitted(lm_mecaexpo_np)
residuals_values <- residuals(lm_mecaexpo_np)
# Check Linearity
plot(fitted_values, residuals_values, main="Fitted Values vs Residuals", xlab="Fitted Values", ylab="Residuals")
abline(h=0, col="red") # residuals are randomly scattered around zero
# Normality of residuals
qqnorm(lm_mecaexpo_np$residuals)
qqline(lm_mecaexpo_np$residuals, col = "red")
hist(lm_mecaexpo_np$residuals, main = "Residuals Histogram", xlab = "Residuals", col = "lightgray", border = "black")
# QQ plot shows points that fall approximately along a straight line = Normal distribution
# Homoscedasticity
plot(lm_mecaexpo_np$fitted.values, abs(lm_mecaexpo_np$residuals))
abline(h = 0, col = "red")# points are randomly scattered acros the range of fitted values
#  Independence of Residuals- are the residuals (errors) correlated with each other
dwtest(lm_mecaexpo_np) # no autocorrelation
# multicollinearity
vif(lm_mecaexpo_np)
# Outliers
plot(lm_mecaexpo_np, which = 4)  # Cook's distance plot
# All models meet the assumptions of multiple linear regression 
```
## Standardised livestock variables
```{r}
# The livestock exposure metrics have different scales. Standardizing the variables (i.e., transforming them to have a mean of 0 and a standard deviation of 1) allows for fairer comparisons of their coefficients, since each variable will be on the same scale.
# Different Scales: The livestock exposure metrics are measured on different scales (e.g., raw counts of animals, predicted values for microbial agents). Without standardization, variables with larger scales (e.g., counts of livestock) could dominate the model, while variables with smaller scales (e.g., microbial exposure predictions) might be underrepresented.Coefficient Interpretation: When variables are standardized, the coefficients from your regression model can be interpreted as the effect of a one standard deviation change in each predictor variable on the Shannon diversity. This makes it easier to compare the relative importance of each variable (since they are all now on the same scale).In some models, especially when using regularization or more complex algorithms, standardization can help improve numerical stability and convergence.

# NP niche
lm_animals3km_np_std <- lm(Shannon ~  KRD_nCowsWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nPoultryWghtDist.3000m.sum_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_animals3km_np_std$niche <- "NP"

# Loop through the coefficients and add the 'livestock' label based on the variable name
lm_animals3km_np_std$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Cattle (std)", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Goats (std)", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Pigs (std)", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Chickens (std)", NA))))


# modelled microbial agents
lm_endoexpo_np_std <- lm(Shannon ~  EU_ANNUALAVG_PM10_2020_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_np_std$livestock <- "Dispersion-modelled endotoxin (std)"
lm_endoexpo_np_std$niche <- "NP"


lm_ecoliexpo_np_std <- lm(Shannon ~  ecoli_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_np_std$livestock <- "RF-modelled E. coli (std)"
lm_ecoliexpo_np_std$niche <- "NP"


lm_staphexpo_np_std <- lm(Shannon ~  staph_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_np_std$livestock <- "RF-modelled Staph (std)"
lm_staphexpo_np_std$niche <- "NP"


lm_tetwexpo_np_std<- lm(Shannon ~  tetw_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_np_std$livestock <- "RF-modelled tetW (std)"
lm_tetwexpo_np_std$niche <- "NP"

lm_mecaexpo_np_std <- lm(Shannon ~  meca_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_np_std$livestock <- "RF-modelled mecA (std)"
lm_mecaexpo_np_std$niche <- "NP"


# OP niche
lm_animals3km_op_std <- lm(Shannon ~  KRD_nCowsWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nPoultryWghtDist.3000m.sum_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp) %>% 
  broom::tidy(., conf.int = T)
lm_animals3km_op_std$niche <- "OP"

lm_animals3km_op_std$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Cattle (std)", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Goats (std)", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Pigs (std)", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Chickens (std)", NA))))


# modelled microbial agents
lm_endoexpo_op_std <- lm(Shannon ~  EU_ANNUALAVG_PM10_2020_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_op_std$livestock <- "Dispersion-modelled endotoxin (std)"
lm_endoexpo_op_std$niche <- "OP"

lm_ecoliexpo_op_std <- lm(Shannon ~  ecoli_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_op_std$livestock <- "RF-modelled E. coli (std)"
lm_ecoliexpo_op_std$niche <- "OP"

lm_staphexpo_op_std <- lm(Shannon ~  staph_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_op_std$livestock <- "RF-modelled Staph (std)"
lm_staphexpo_op_std$niche <- "OP"

lm_tetwexpo_op_std <- lm(Shannon ~  tetw_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_op_std$livestock <- "RF-modelled tetW (std)"
lm_tetwexpo_op_std$niche <- "OP"

lm_mecaexpo_op_std <- lm(Shannon ~  meca_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_op_std$livestock <- "RF-modelled mecA (std)"
lm_mecaexpo_op_std$niche <- "OP"


# Combine all the tidy data
combined_data_shannon <- bind_rows(lm_animals3km_np_std, lm_endoexpo_np_std, lm_ecoliexpo_np_std, lm_staphexpo_np_std, lm_tetwexpo_np_std, lm_mecaexpo_np_std, lm_animals3km_op_std, lm_endoexpo_op_std, lm_ecoliexpo_op_std, lm_staphexpo_op_std, lm_tetwexpo_op_std, lm_mecaexpo_op_std)

# Filter to include only livestock exposure terms
livestock_terms_shannon <- combined_data_shannon %>%
  filter(term %in% c("KRD_nCowsWghtDist.3000m.sum_std", 
                     "KRD_nGoatsWghtDist.3000m.sum_std", 
                     "KRD_nPigsWghtDist.3000m.sum_std", 
                     "KRD_nPoultryWghtDist.3000m.sum_std",
                     "EU_ANNUALAVG_PM10_2020_std",
                     "ecoli_RF_preds_std",
                     "staph_RF_preds_std",
                     "meca_RF_preds_std",
                     "tetw_RF_preds_std"))

# Apply multiple testing correction (Benjamini-Hochberg) for NP
livestock_terms_np_shannon <- livestock_terms_shannon %>%
  filter(niche == "NP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Apply multiple testing correction (Benjamini-Hochberg) for OP
livestock_terms_op_shannon <- livestock_terms_shannon %>%
  filter(niche == "OP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))
    
# Combine the results for NP and OP niches
livestock_terms_corrected_shannon <- bind_rows(livestock_terms_np_shannon, livestock_terms_op_shannon)

# Add adjusted p-value labels
livestock_terms_corrected_shannon <- livestock_terms_corrected_shannon %>%
  mutate(
    label = case_when(
      p.adjusted < 0.1 ~ "*",  # If p.adjusted is less than 0.1, assign *
      TRUE ~ NA_character_     # Otherwise, set label to NA
    )
  )
write.csv(livestock_terms_corrected_shannon,"../Output/Alpha_diversity/Residential_livestock_expo_shannondiversity_linear_regression_std.csv")


# Define the desired order of the variables and their new names
livestock_terms_corrected_shannon$livestock <- factor(
  livestock_terms_corrected_shannon$livestock,
  levels = c(
    "Cattle (std)",
    "Chickens (std)",
    "Goats (std)",
    "Pigs (std)",
    "RF-modelled mecA (std)",
    "RF-modelled tetW (std)",
    "RF-modelled Staph (std)",
    "RF-modelled E. coli (std)",
    "Dispersion-modelled endotoxin (std)"
  )
)

# Shannon
livestock_expo_shannon_std <- ggplot(livestock_terms_corrected_shannon, aes(x = livestock, y = estimate, color = niche)) +
  geom_point(position = position_dodge(width = 0.5), size = 3) +  # Plot points
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.2, 
                position = position_dodge(width = 0.5)) +  # Error bars for confidence intervals
  geom_text(aes(label = label), 
            position = position_dodge(width = 0.5), 
            vjust = -0.5, size = 4, color = "black") +  # Add labels above the points
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +  # Add dotted line at 0
  labs(x = "Livestock exposure (standardised)", 
       y = "Shannon Diversity Regression Coefficient",
       color = "Niche") +
  theme_minimal(base_size = 16) +  # Increase base font size
  theme(
    legend.position = "top",  # Move legend to the top
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),  # Increase size of axis titles
    axis.text = element_text(size = 12),   # Increase size of axis text
    plot.title = element_text(size = 18)   # Increase size of the plot title
  ) +
  scale_color_manual(values = c("NP" = "#dd968e", "OP" = "#547484")) +  # Custom colors for NP and OP
  facet_wrap(~ niche, scales = "free_x") +  # Facet by niche (NP and OP)
  coord_flip()   # Flip coordinates for better readability
livestock_expo_shannon_std

# Save the plot
theme(
  plot.margin = margin(1, 1, 2, 1, "cm")  # More space at the bottom (2 cm)
)

ggsave("../Output/Alpha_diversity/livestock_expo_shannon_standardised.svg", livestock_expo_shannon_std, width = 8, height = 5)




# OBSERVED RICHNESS
# NP niche
lm_animals3km_np_std <- lm(Observed ~  KRD_nCowsWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nPoultryWghtDist.3000m.sum_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_animals3km_np_std$niche <- "NP"

# Loop through the coefficients and add the 'livestock' label based on the variable name
lm_animals3km_np_std$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Cattle (std)", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Goats (std)", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Pigs (std)", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Chickens (std)", NA))))


# modelled microbial agents
lm_endoexpo_np_std <- lm(Observed ~  EU_ANNUALAVG_PM10_2020_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_np_std$livestock <- "Dispersion-modelled endotoxin (std)"
lm_endoexpo_np_std$niche <- "NP"


lm_ecoliexpo_np_std <- lm(Observed ~  ecoli_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_np_std$livestock <- "RF-modelled E. coli (std)"
lm_ecoliexpo_np_std$niche <- "NP"


lm_staphexpo_np_std <- lm(Observed ~  staph_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_np_std$livestock <- "RF-modelled Staph (std)"
lm_staphexpo_np_std$niche <- "NP"


lm_tetwexpo_np_std<- lm(Observed ~  tetw_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_np_std$livestock <- "RF-modelled tetW (std)"
lm_tetwexpo_np_std$niche <- "NP"

lm_mecaexpo_np_std <- lm(Observed ~  meca_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_np_std$livestock <- "RF-modelled mecA (std)"
lm_mecaexpo_np_std$niche <- "NP"


# OP niche
lm_animals3km_op_std <- lm(Observed ~  KRD_nCowsWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nPoultryWghtDist.3000m.sum_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp) %>% 
  broom::tidy(., conf.int = T)
lm_animals3km_op_std$niche <- "OP"

lm_animals3km_op_std$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Cattle (std)", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Goats (std)", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Pigs (std)", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Chickens (std)", NA))))


# modelled microbial agents
lm_endoexpo_op_std <- lm(Observed ~  EU_ANNUALAVG_PM10_2020_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_op_std$livestock <- "Dispersion-modelled endotoxin (std)"
lm_endoexpo_op_std$niche <- "OP"

lm_ecoliexpo_op_std <- lm(Observed ~  ecoli_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_op_std$livestock <- "RF-modelled E. coli (std)"
lm_ecoliexpo_op_std$niche <- "OP"

lm_staphexpo_op_std <- lm(Observed ~  staph_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_op_std$livestock <- "RF-modelled Staph (std)"
lm_staphexpo_op_std$niche <- "OP"

lm_tetwexpo_op_std <- lm(Observed ~  tetw_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_op_std$livestock <- "RF-modelled tetW (std)"
lm_tetwexpo_op_std$niche <- "OP"

lm_mecaexpo_op_std <- lm(Observed ~  meca_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_op_std$livestock <- "RF-modelled mecA (std)"
lm_mecaexpo_op_std$niche <- "OP"


# Combine all the tidy data
combined_data_Observed <- bind_rows(lm_animals3km_np_std, lm_endoexpo_np_std, lm_ecoliexpo_np_std, lm_staphexpo_np_std, lm_tetwexpo_np_std, lm_mecaexpo_np_std, lm_animals3km_op_std, lm_endoexpo_op_std, lm_ecoliexpo_op_std, lm_staphexpo_op_std, lm_tetwexpo_op_std, lm_mecaexpo_op_std)

# Filter to include only livestock exposure terms
livestock_terms_Observed <- combined_data_Observed %>%
  filter(term %in% c("KRD_nCowsWghtDist.3000m.sum_std", 
                     "KRD_nGoatsWghtDist.3000m.sum_std", 
                     "KRD_nPigsWghtDist.3000m.sum_std", 
                     "KRD_nPoultryWghtDist.3000m.sum_std",
                     "EU_ANNUALAVG_PM10_2020_std",
                     "ecoli_RF_preds_std",
                     "staph_RF_preds_std",
                     "meca_RF_preds_std",
                     "tetw_RF_preds_std"))

# Apply multiple testing correction (Benjamini-Hochberg) for NP
livestock_terms_np_Observed <- livestock_terms_Observed %>%
  filter(niche == "NP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Apply multiple testing correction (Benjamini-Hochberg) for OP
livestock_terms_op_Observed <- livestock_terms_Observed %>%
  filter(niche == "OP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))
    
# Combine the results for NP and OP niches
livestock_terms_corrected_Observed <- bind_rows(livestock_terms_np_Observed, livestock_terms_op_Observed)

# Add adjusted p-value labels
livestock_terms_corrected_Observed <- livestock_terms_corrected_Observed %>%
  mutate(
    label = case_when(
      p.adjusted < 0.1 ~ "*",  # If p.adjusted is less than 0.1, assign *
      TRUE ~ NA_character_     # Otherwise, set label to NA
    )
  )
write.csv(livestock_terms_corrected_Observed,"../Output/Alpha_diversity/Residential_livestock_expo_Observeddiversity_linear_regression_std.csv")


# Define the desired order of the variables and their new names
livestock_terms_corrected_Observed$livestock <- factor(
  livestock_terms_corrected_Observed$livestock,
  levels = c(
    "Cattle (std)",
    "Chickens (std)",
    "Goats (std)",
    "Pigs (std)",
    "RF-modelled mecA (std)",
    "RF-modelled tetW (std)",
    "RF-modelled Staph (std)",
    "RF-modelled E. coli (std)",
    "Dispersion-modelled endotoxin (std)"
  )
)

# Observed
livestock_expo_Observed_std <- ggplot(livestock_terms_corrected_Observed, aes(x = livestock, y = estimate, color = niche)) +
  geom_point(position = position_dodge(width = 0.5), size = 3) +  # Plot points
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.2, 
                position = position_dodge(width = 0.5)) +  # Error bars for confidence intervals
  geom_text(aes(label = label), 
            position = position_dodge(width = 0.5), 
            vjust = -0.5, size = 4, color = "black") +  # Add labels above the points
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +  # Add dotted line at 0
  labs(x = "Livestock exposure (standardised)", 
       y = "Observed Diversity Regression Coefficient",
       color = "Niche") +
  theme_minimal(base_size = 16) +  # Increase base font size
  theme(
    legend.position = "top",  # Move legend to the top
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),  # Increase size of axis titles
    axis.text = element_text(size = 12),   # Increase size of axis text
    plot.title = element_text(size = 18)   # Increase size of the plot title
  ) +
  scale_color_manual(values = c("NP" = "#dd968e", "OP" = "#547484")) +  # Custom colors for NP and OP
  facet_wrap(~ niche, scales = "free_x") +  # Facet by niche (NP and OP)
  coord_flip()   # Flip coordinates for better readability
livestock_expo_Observed_std

# Save the plot
theme(
  plot.margin = margin(1, 1, 2, 1, "cm")  # More space at the bottom (2 cm)
)

ggsave("../Output/Alpha_diversity/livestock_expo_Observed_standardised.svg", livestock_expo_Observed_std, width = 8, height = 5)






# qPCR
# NP niche
lm_animals3km_np_std <- lm(qPCR ~  KRD_nCowsWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nPoultryWghtDist.3000m.sum_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_animals3km_np_std$niche <- "NP"

# Loop through the coefficients and add the 'livestock' label based on the variable name
lm_animals3km_np_std$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Cattle (std)", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Goats (std)", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Pigs (std)", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum_std", lm_animals3km_np_std$term), "Chickens (std)", NA))))


# modelled microbial agents
lm_endoexpo_np_std <- lm(qPCR ~  EU_ANNUALAVG_PM10_2020_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_np_std$livestock <- "Dispersion-modelled endotoxin (std)"
lm_endoexpo_np_std$niche <- "NP"


lm_ecoliexpo_np_std <- lm(qPCR ~  ecoli_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_np_std$livestock <- "RF-modelled E. coli (std)"
lm_ecoliexpo_np_std$niche <- "NP"


lm_staphexpo_np_std <- lm(qPCR ~  staph_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_np_std$livestock <- "RF-modelled Staph (std)"
lm_staphexpo_np_std$niche <- "NP"


lm_tetwexpo_np_std<- lm(qPCR ~  tetw_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_np_std$livestock <- "RF-modelled tetW (std)"
lm_tetwexpo_np_std$niche <- "NP"

lm_mecaexpo_np_std <- lm(qPCR ~  meca_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_np_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_np_std$livestock <- "RF-modelled mecA (std)"
lm_mecaexpo_np_std$niche <- "NP"


# OP niche
lm_animals3km_op_std <- lm(qPCR ~  KRD_nCowsWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nPoultryWghtDist.3000m.sum_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp) %>% 
  broom::tidy(., conf.int = T)
lm_animals3km_op_std$niche <- "OP"

lm_animals3km_op_std$livestock <- ifelse(grepl("KRD_nCowsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Cattle (std)", 
                                     ifelse(grepl("KRD_nGoatsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Goats (std)", 
                                            ifelse(grepl("KRD_nPigsWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Pigs (std)", 
                                                   ifelse(grepl("KRD_nPoultryWghtDist.3000m.sum_std", lm_animals3km_op_std$term), "Chickens (std)", NA))))


# modelled microbial agents
lm_endoexpo_op_std <- lm(qPCR ~  EU_ANNUALAVG_PM10_2020_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_endoexpo_op_std$livestock <- "Dispersion-modelled endotoxin (std)"
lm_endoexpo_op_std$niche <- "OP"

lm_ecoliexpo_op_std <- lm(qPCR ~  ecoli_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_ecoliexpo_op_std$livestock <- "RF-modelled E. coli (std)"
lm_ecoliexpo_op_std$niche <- "OP"

lm_staphexpo_op_std <- lm(qPCR ~  staph_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_staphexpo_op_std$livestock <- "RF-modelled Staph (std)"
lm_staphexpo_op_std$niche <- "OP"

lm_tetwexpo_op_std <- lm(qPCR ~  tetw_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_tetwexpo_op_std$livestock <- "RF-modelled tetW (std)"
lm_tetwexpo_op_std$niche <- "OP"

lm_mecaexpo_op_std <- lm(qPCR ~  meca_RF_preds_std + age + sampling_season + gender + smoked_ever, data = meta_op_cp)%>% 
  broom::tidy(., conf.int = T)
lm_mecaexpo_op_std$livestock <- "RF-modelled mecA (std)"
lm_mecaexpo_op_std$niche <- "OP"


# Combine all the tidy data
combined_data_qPCR_std <- bind_rows(lm_animals3km_np_std, lm_endoexpo_np_std, lm_ecoliexpo_np_std, lm_staphexpo_np_std, lm_tetwexpo_np_std, lm_mecaexpo_np_std, lm_animals3km_op_std, lm_endoexpo_op_std, lm_ecoliexpo_op_std, lm_staphexpo_op_std, lm_tetwexpo_op_std, lm_mecaexpo_op_std)

# Filter to include only livestock exposure terms
livestock_terms_qPCR_std <- combined_data_qPCR_std %>%
  filter(term %in% c("KRD_nCowsWghtDist.3000m.sum_std", 
                     "KRD_nGoatsWghtDist.3000m.sum_std", 
                     "KRD_nPigsWghtDist.3000m.sum_std", 
                     "KRD_nPoultryWghtDist.3000m.sum_std",
                     "EU_ANNUALAVG_PM10_2020_std",
                     "ecoli_RF_preds_std",
                     "staph_RF_preds_std",
                     "meca_RF_preds_std",
                     "tetw_RF_preds_std"))

# Apply multiple testing correction (Benjamini-Hochberg) for NP
livestock_terms_np_qPCR_std <- livestock_terms_qPCR_std %>%
  filter(niche == "NP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))

# Apply multiple testing correction (Benjamini-Hochberg) for OP
livestock_terms_op_qPCR_std <- livestock_terms_qPCR_std %>%
  filter(niche == "OP") %>%
  mutate(p.adjusted = p.adjust(p.value, method = "BH"))
    
# Combine the results for NP and OP niches
livestock_terms_corrected_qPCR_std <- bind_rows(livestock_terms_np_qPCR_std, livestock_terms_op_qPCR_std)

# Add adjusted p-value labels
livestock_terms_corrected_qPCR_std <- livestock_terms_corrected_qPCR_std %>%
  mutate(
    label = case_when(
      p.adjusted < 0.1 ~ "*",  # If p.adjusted is less than 0.1, assign *
      TRUE ~ NA_character_     # Otherwise, set label to NA
    )
  )
write.csv(livestock_terms_corrected_qPCR_std,"../Output/Alpha_diversity/Residential_livestock_expo_qPCRdiversity_linear_regression_std.csv")


# Define the desired order of the variables and their new names
livestock_terms_corrected_qPCR_std$livestock <- factor(
  livestock_terms_corrected_qPCR_std$livestock,
  levels = c(
    "Cattle (std)",
    "Chickens (std)",
    "Goats (std)",
    "Pigs (std)",
    "RF-modelled mecA (std)",
    "RF-modelled tetW (std)",
    "RF-modelled Staph (std)",
    "RF-modelled E. coli (std)",
    "Dispersion-modelled endotoxin (std)"
  )
)

# qPCR
livestock_expo_qPCR_std <- ggplot(livestock_terms_corrected_qPCR_std, aes(x = livestock, y = estimate, color = niche)) +
  geom_point(position = position_dodge(width = 0.5), size = 3) +  # Plot points
  geom_errorbar(aes(ymin = conf.low, ymax = conf.high), width = 0.2, 
                position = position_dodge(width = 0.5)) +  # Error bars for confidence intervals
    geom_text(aes(label = label), 
            position = position_dodge(width = 0.5), 
            vjust = -0.5, size = 4, color = "black") +  # Add labels above the points
  geom_hline(yintercept = 0, linetype = "dotted", color = "black") +  # Add dotted line at 0
  labs(x = "Livestock exposure (std)", 
       y = "qPCR Regression Coefficient",
       color = "Niche") +
  theme_minimal(base_size = 16) +  # Increase base font size
  theme(
    legend.position = "top",  # Move legend to the top
    panel.border = element_rect(color = "black", fill = NA),  # Add black borders around the facets
    strip.text = element_text(colour = 'white', face = 'bold', size = 16),  # Increase size for facet titles
    strip.background = element_rect(fill = "#2F4858", color = "#2F4858"),  # Color facet box and border
    axis.title = element_text(size = 14),  # Increase size of axis titles
    axis.text = element_text(size = 12),   # Increase size of axis text
    plot.title = element_text(size = 18)   # Increase size of the plot title
  ) +
  scale_color_manual(values = c("NP" = "#dd968e", "OP" = "#547484")) +  # Custom colors for NP and OP
  facet_wrap(~ niche, scales = "free_x") +  # Facet by niche (NP and OP)
  coord_flip()   # Flip coordinates for better readability
livestock_expo_qPCR_std

# Save the plot
theme(
  plot.margin = margin(1, 1, 2, 1, "cm")  # More space at the bottom (2 cm)
)

ggsave("../Output/Alpha_diversity/livestock_expo_qPCR_standardised.svg", livestock_expo_qPCR_std, width = 8, height = 5)


```
