==> Scripts/7b_Multivariable_PERMANOVAs_PCoAs.Rmd <==
---
title: "Beta diversity"
author: "Beatrice Cornu Hewitt" 
date: "`r Sys.time()`"
---

# Packages
```{r}
library(tidyverse);library(magrittr);library(glue);library(here);library(ggtext);library(fs);library(phyloseq);library(microbiomer);library(vegan);library(microbiome);library(decontam);library(reshape2);library(colorspace);library(scales);library(ggpubr);library(ggvenn);library(ggnewscale);library(RColorBrewer);library(paletteer);library(broom);library(rcartocolor);library(microViz);library(ggside);library(emmeans); library(MatchIt); library(glue); library(readxl)

# Set theme for plots in notebook
theme_set(theme_bw())
theme_update(axis.text.x = element_text(angle = 45, hjust=1),
             strip.text = element_text(colour = 'white'),
             strip.background =element_rect(fill="#2F4858", color = "#2F4858"))

source("functions.R")

```

# Data
```{r}
ps <- readRDS("../Output/Phyloseq/2_ps_complete.Rds")
# the phyloseq was already rarefied (ps2 <- phyloseq::rarefy_even_depth(ps1, replace = T, trimOTUs = T)), and the alpha diversity indices were computed from this rarefied data in the script '2b_clean_phyloseq.Rmd': 
# adiv <- data.frame(
#   "Observed" = phyloseq::estimate_richness(ps2, measures = "Observed"),
#   "Shannon" = phyloseq::estimate_richness(ps2, measures = "Shannon"))

# Read in the metadata variable names and explanations
names <- readxl::read_xlsx("../Input/Codebook_metadata_all.xlsx")
```
# Functions for multivariable PERMANOVA
```{r}
beta_df <- function(ord, meta, var) { 
  m_sub <- subset(meta, !is.na(meta[var]))
  m_sub <- subset(m_sub, !is.na(m_sub["age"]))
  m_sub <- subset(m_sub, !is.na(m_sub["sampling_season"]))  
  m_sub <- subset(m_sub, !is.na(m_sub["gender"]))
  m_sub <- subset(m_sub, !is.na(m_sub["smoked_ever"]))
  
  ord_sub <- ord[rownames(ord) %in% m_sub$sample_id, colnames(ord) %in% 
                   m_sub$sample_id]
  ord_sub <- as.dist(ord_sub)
  return(list(ord_sub = ord_sub, m_sub = m_sub))
}

do_permanova <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever")), 
               permutations = 999, data = meta)
  return(t)
}


do_permanova_com <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA)  # Now all p-values < 0.1 get "*"
  )

  return(t)
}


```
# Filter phyloseq and convert to relative abundance 
```{r}
# Filter the ps object by niche and population
# Apply a prevalence/abundance filter to the phyloseq objects
# Then convert to RA
ps_np <- subset_samples(ps, niche == "NP") %>% pres_abund_filter() %>% to_RA() 
ps_op <- subset_samples(ps, niche == "OP") %>% pres_abund_filter() %>% to_RA() 
ps_cp <- subset_samples(ps, population == "CP") %>% pres_abund_filter() %>% to_RA() 
ps_gf <- subset_samples(ps, population == "GF") %>% pres_abund_filter() %>% to_RA()  
ps_all <- ps %>% pres_abund_filter() %>% to_RA() 

# Convert the otu tables from the above phyloseq objects to dataframes
otu_np <- as.data.frame(otu_table(ps_np)) %>% t()
otu_op <- as.data.frame(otu_table(ps_op)) %>% t()
otu_cp <- as.data.frame(otu_table(ps_cp)) %>% t()
otu_gf <- as.data.frame(otu_table(ps_gf)) %>% t()
otu_all <- as.data.frame(otu_table(ps_all)) %>% t()

# Extract metadata for each ps object
df_np <- meta_to_df(ps_np)
df_op <- meta_to_df(ps_op)
df_cp <- meta_to_df(ps_cp)
df_gf <- meta_to_df(ps_gf)
df_all <- meta_to_df(ps_all)

# Compute BC dissimilarity matrices for each ps
ord_np <- vegdist(otu_np, method = "bray") %>% as.matrix()
ord_op <- vegdist(otu_op, method = "bray") %>% as.matrix()
ord_cp <- vegdist(otu_cp, method = "bray") %>% as.matrix()
ord_gf <- vegdist(otu_gf, method = "bray") %>% as.matrix()
ord_all <- vegdist(otu_all, method = "bray") %>% as.matrix()

# Compute mean BC dissimilarity
mean(ord_np)
mean(ord_op)
```
# Niche (NP vs OP) - multivariable PERMANOVA
```{r}
# Niche effect per population (corrected for age + sampling_season + gender + smoked_ever)
niche_OPvsNP_CP <- do_permanova_com(ord_cp, df_cp, "niche") %>% unnest()
niche_OPvsNP_CP$group <- "CP"

niche_OPvsNP_GF <- do_permanova_com(ord_gf, df_gf, "niche") %>% unnest()
niche_OPvsNP_GF$group <- "GF"


# Niche effect for both populations together (corrected for age + sampling_season + gender + smoked_ever)
niche_OPvsNP <- do_permanova_com(ord_all, df_all, "niche") %>% unnest()
niche_OPvsNP$group <- "Both populations"

```

# PCoAs
## Niche for CP
```{r}
# Extract R2 and p-value from multivariable PERMANOVA
niche_CP_r2 <- round(niche_OPvsNP_CP$R2, 3)
niche_CP_r2_pval <- niche_OPvsNP_CP$p.value
niche_CP_r2_sig <- niche_OPvsNP_CP$label

PCoA_niche_CP <- ps_cp %>% 
  dist_calc("bray") %>%
  ord_calc(method = "PCoA") %>%
  ord_plot(color = "niche", shape = "niche", size = 3.5, alpha = 0.7) +
  scale_color_manual(values = c("#dd968e","#547484"), aesthetics = c("colour")) +
  scale_fill_manual(values = c("#dd968e","#547484"), aesthetics = c("fill")) +
  stat_ellipse(aes(fill = niche, color = niche), 
               type = "t", level = 0.95, size = 1,geom = "polygon", alpha = 0.1) +
  theme_bw() +
  ggside::geom_xsideboxplot(aes(fill = niche, y = niche), orientation = "y") +
  ggside::geom_ysideboxplot(aes(fill = niche, x = niche), orientation = "x") +
  ggside::scale_xsidey_discrete(labels = NULL) +
  ggside::scale_ysidex_discrete(labels = NULL) +
  ggside::theme_ggside_void()+
  # Add PERMANOVA R² and p-value as annotation
annotate("text", x = Inf, y = -2.2, 
           label = paste0("R² = ", niche_CP_r2, "\n", 
                          "p value = ", format(niche_CP_r2_pval, scientific = TRUE), " ", niche_CP_r2_sig), 
           hjust = 1.1, vjust = 1, size = 4, color = "black")

PCoA_niche_CP
ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/PCoA_niche_CP.svg", PCoA_niche_CP, width= 5.5, height=5)
```
## Niche for CP and GF together
```{r}
# Extract R2 and p-value from multivariable PERMANOVA
niche_r2 <- round(niche_OPvsNP$R2, 3)
niche_r2_pval <- niche_OPvsNP$p.value
niche_r2_sig <- niche_OPvsNP$label

PCoA_niche <- ps_all %>% 
  dist_calc("bray") %>%
  ord_calc(method = "PCoA") %>%
  ord_plot(color = "niche", shape = "niche", size = 3.5, alpha = 0.7) +
  scale_color_manual(values = c("#dd968e","#547484"), aesthetics = c("colour")) +
  scale_fill_manual(values = c("#dd968e","#547484"), aesthetics = c("fill")) +
  stat_ellipse(aes(fill = niche, color = niche), 
               type = "t", level = 0.95, size = 1,geom = "polygon", alpha = 0.1) +
  theme_bw() +
  ggside::geom_xsideboxplot(aes(fill = niche, y = niche), orientation = "y") +
  ggside::geom_ysideboxplot(aes(fill = niche, x = niche), orientation = "x") +
  ggside::scale_xsidey_discrete(labels = NULL) +
  ggside::scale_ysidex_discrete(labels = NULL) +
  ggside::theme_ggside_void()+
  # Add PERMANOVA R² and p-value as annotation
annotate("text", x = Inf, y = -2.2, 
           label = paste0("R² = ", niche_r2, "\n", 
                          "p value = ", format(niche_r2_pval, scientific = TRUE), " ", niche_r2_sig), 
           hjust = 1.1, vjust = 1, size = 4, color = "black")

PCoA_niche
ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/PCoA_niche.svg", PCoA_niche, width= 5.5, height=5)
```


# Population multivariable model
```{r}
ps_cp_gf_matched <- readRDS("../Output/Phyloseq/3_ps_complete_matchedCPGF.Rds")

# Apply a prevalence/abundance filter to the phyloseq objects
# Then convert to RA
ps_np_cp_gf_matched <- subset_samples(ps_cp_gf_matched, niche == "NP") %>% pres_abund_filter() %>% to_RA() 
ps_op_cp_gf_matched <- subset_samples(ps_cp_gf_matched, niche == "OP") %>% pres_abund_filter() %>% to_RA() 

# Convert the otu tables from the above phyloseq objects to dataframes
otu_np_cp_gf_matched <- as.data.frame(otu_table(ps_np_cp_gf_matched)) %>% t()
otu_op_cp_gf_matched <- as.data.frame(otu_table(ps_op_cp_gf_matched)) %>% t()

# Extract metadata for each ps object
df_np_cp_gf_matched <- meta_to_df(ps_np_cp_gf_matched)
df_op_cp_gf_matched <- meta_to_df(ps_op_cp_gf_matched)

# Compute BC dissimilarity matrices for each ps
ord_np_cp_gf_matched <- vegdist(otu_np_cp_gf_matched, method = "bray") %>% as.matrix()
ord_op_cp_gf_matched <- vegdist(otu_op_cp_gf_matched, method = "bray") %>% as.matrix()

# Population effect (corrected for age + sampling_season + gender + smoked_ever)
population_CPvsGF_NP <- do_permanova_com(ord_np_cp_gf_matched, df_np_cp_gf_matched, "population") %>% unnest()
population_CPvsGF_NP$group <- "NP"

population_CPvsGF_OP <- do_permanova_com(ord_op_cp_gf_matched, df_op_cp_gf_matched, "population") %>% unnest()
population_CPvsGF_OP$group <- "OP"

# Combine mutivariable PERMANOVA results for population (both NP and OP)
population_permanova <- rbind(population_CPvsGF_NP, population_CPvsGF_OP)

# save as csv
write.csv(population_permanova, "../Output/Beta_diversity/Multivariable_PERMANOVAs/Occupational_CP_GF_PERMANOVA.csv")
```

# PCoA plots
## CP vs GF for NP niche
```{r}
ps_np_cp_gf_matched@sam_data$population <- factor(ps_np_cp_gf_matched@sam_data$population, levels = c("CP", "GF"))
# Extract R2 and p-value from multivariable PERMANOVA
pop_NP_r2 <- round(population_CPvsGF_NP$R2, 3)
pop_NP_r2_pval <- population_CPvsGF_NP$p.value
pop_NP_r2_sig <- population_CPvsGF_NP$label


PCoA_pop_NP <- ps_np_cp_gf_matched %>% 
  dist_calc("bray") %>%
  ord_calc(method = "PCoA") %>%
  ord_plot(color = "population", shape = "population", size = 3.5, alpha = 0.7) + # Adjust transparency for points
  scale_color_manual(values = c("#e3bd6a", "#76b6a4"), aesthetics = c("colour")) +
  scale_fill_manual(values = c("#e3bd6a", "#76b6a4"), aesthetics = c("fill")) +
  stat_ellipse(aes(fill = population, color = population), 
               type = "t", level = 0.95, size = 1, geom = "polygon", alpha = 0.1) + # Transparent fill inside ellipses
  theme_bw() +
  theme(
    axis.title = element_text(size = 14),  # Increase axis title size
    axis.text = element_text(size = 12)    # Increase axis text size
  ) +
  ggside::geom_xsideboxplot(aes(fill = population, y = population), orientation = "y") +
  ggside::geom_ysideboxplot(aes(fill = population, x = population), orientation = "x") +
  ggside::scale_xsidey_discrete(labels = NULL) +
  ggside::scale_ysidex_discrete(labels = NULL) +
  ggside::theme_ggside_void() +
  # Add PERMANOVA R² and p-value as annotation
  annotate("text", x = Inf, y = -1.3, 
           label = paste0("R² = ", pop_NP_r2, "\n", 
                          "p value = ", format(pop_NP_r2_pval, scientific = TRUE), " ", pop_NP_r2_sig), 
           hjust = 1.1, vjust = 1.5, size = 4, color = "black")

PCoA_pop_NP

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/PCoA_GFvsCP_NP.svg", PCoA_pop_NP, width= 6, height=5)
```
## CP vs GF for OP niche
```{r}
ps_op_cp_gf_matched@sam_data$population <- factor(ps_op_cp_gf_matched@sam_data$population, levels = c("CP", "GF"))
# Extract R2 and p-value from multivariable PERMANOVA
pop_OP_r2 <- round(population_CPvsGF_OP$R2, 3)
pop_OP_r2_pval <- population_CPvsGF_OP$p.value
pop_OP_r2_sig <- population_CPvsGF_OP$label


PCoA_pop_op <- ps_op_cp_gf_matched %>% 
  dist_calc("bray") %>%
  ord_calc(method = "PCoA") %>%
  ord_plot(color = "population", shape = "population", size = 3.5, alpha = 0.7) + # Adjust transparency for points
  scale_color_manual(values = c("#e3bd6a", "#76b6a4"), aesthetics = c("colour")) +
  scale_fill_manual(values = c("#e3bd6a", "#76b6a4"), aesthetics = c("fill")) +
  stat_ellipse(aes(fill = population, color = population), 
               type = "t", level = 0.95, size = 1, geom = "polygon", alpha = 0.1) + # Transparent fill inside ellipses
  theme_bw() +
  theme(
    axis.title = element_text(size = 14),  # Increase axis title size
    axis.text = element_text(size = 12)    # Increase axis text size
  ) +
  ggside::geom_xsideboxplot(aes(fill = population, y = population), orientation = "y") +
  ggside::geom_ysideboxplot(aes(fill = population, x = population), orientation = "x") +
  ggside::scale_xsidey_discrete(labels = NULL) +
  ggside::scale_ysidex_discrete(labels = NULL) +
  ggside::theme_ggside_void() +
  # Add PERMANOVA R² and p-value as annotation
  annotate("text", x = Inf, y = -1.3, 
           label = paste0("R² = ", pop_OP_r2, "\n", 
                          "p value = ", format(pop_OP_r2_pval, scientific = TRUE), " ", pop_OP_r2_sig), 
           hjust = 1.1, vjust = 1.5, size = 4, color = "black")

PCoA_pop_op

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/PCoA_GFvsCP_OP.svg", PCoA_pop_op, width= 6, height=5)
```

## GF OP niche - different farms coloured 
```{r}
ps_op_gf <- subset_samples(ps_op_cp_gf_matched, population == "GF") %>% pres_abund_filter() %>% to_RA()  
# A total of 351 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 319 ASVs excluded).
# 92 samples
getwd()
farm_data <- as.data.frame(read_excel("../Input/FarmerID_farmID.xlsx"))
farm_data$VGO3_ID <- as.character(farm_data$VGO3_ID)

# Add row names as a new column
sample_data_df <- data.frame(sample_data(ps_op_gf))
sample_data_df$Sample_ID <- rownames(sample_data_df)

# Perform left_join, ensuring VGO3_ID matches
sample_data_updated <- sample_data_df %>%
  left_join(farm_data, by = "VGO3_ID")

# Restore row names and remove the temporary Sample_ID column
rownames(sample_data_updated) <- sample_data_updated$Sample_ID
sample_data_updated$Sample_ID <- NULL


sample_data(ps_op_gf) <- sample_data_updated
farm_palette <- colorRampPalette(brewer.pal(12, "Paired"))(40)

sample_data(ps_op_gf)$Farm_ID <- as.factor(sample_data(ps_op_gf)$Farm_ID)

PCoA_gf_op_farm <- ps_op_gf %>% 
  dist_calc("bray") %>%
  ord_calc(method = "PCoA") %>%
  ord_plot(color = "Farm_ID", size = 3.5, alpha = 0.7) + 
  scale_color_hue(aesthetics = c("colour")) +  # Automatically assign colors for points
  scale_fill_hue(aesthetics = c("fill")) +    # Automatically assign colors for fills
  theme_bw() +
  theme(
    axis.title = element_text(size = 14),  # Increase axis title size
    axis.text = element_text(size = 12)    # Increase axis text size
  )

# Show the plot
PCoA_gf_op_farm


ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/PCoA_GF_OP_farm_id.svg", PCoA_gf_op_farm, width= 6, height=5)


# hard to tell with colours, only incude those with non-unique farm IDs 
# Get the 'Farm_ID' from the sample data
# Drop unused factor levels for Farm_ID
sample_data(ps_op_gf_filtered)$Farm_ID <- droplevels(sample_data(ps_op_gf_filtered)$Farm_ID)

# Check the factor levels again
levels(sample_data(ps_op_gf_filtered)$Farm_ID)


# Now, ps_op_gf_filtered contains only samples with 'Farm_ID' values that appear more than once
PCoA_gf_op_farm_filt <- ps_op_gf_filtered %>% 
  dist_calc("bray") %>%
  ord_calc(method = "PCoA") %>%
  ord_plot(color = "Farm_ID", size = 3.5, alpha = 0.7) + 
  scale_color_hue(aesthetics = c("colour")) +  # Automatically assign colors for points
  scale_fill_hue(aesthetics = c("fill")) +    # Automatically assign colors for fills
  theme_bw() +
  theme(
    axis.title = element_text(size = 14),  # Increase axis title size
    axis.text = element_text(size = 12)    # Increase axis text size
  )+
  stat_ellipse(aes(fill = Farm_ID, color = Farm_ID), 
               type = "t", level = 0.95, size = 1, geom = "polygon", alpha = 0.1) # Transparent fill inside ellipses
# Show the plot
PCoA_gf_op_farm_filt

```

# Multivariable PERMANOVAs livestock exposures 
## Prepare data
```{r}
# Filter the ps object by niche and population
# Apply a prevalence/abundance filter to the phyloseq objects
# Then convert to RA
ps_np_cp <- subset_samples(ps, niche == "NP" & population == "CP") %>% 
  pres_abund_filter() %>% 
  to_RA() # A total of 1158 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 19448 ASVs excluded).

ps_op_cp <- subset_samples(ps, niche == "OP" & population == "CP") %>% 
  pres_abund_filter() %>% 
  to_RA() # A total of 1229 ASVs were found to be present at or above a level of confident detection (0.1% relative abundance) in at least 2 samples (n = 19377 ASVs excluded).

# Convert the otu tables from the above phyloseq objects to dataframes
otu_np_cp <- as.data.frame(otu_table(ps_np_cp)) %>% t()
otu_op_cp <- as.data.frame(otu_table(ps_op_cp)) %>% t()

# Extract metadata for each ps object
df_np_cp <- meta_to_df(ps_np_cp)
df_op_cp <- meta_to_df(ps_op_cp)

# Compute BC dissimilarity matrices for each ps
ord_np_cp <- vegdist(otu_np_cp, method = "bray") %>% as.matrix()
ord_op_cp <- vegdist(otu_op_cp, method = "bray") %>% as.matrix()

```
### Non-standardised variables
#### NP 
```{r}
# Microbial exposures
endo_NP_CP <- do_permanova_com(ord_np_cp, df_np_cp, "EU_ANNUALAVG_PM10_2020") %>% unnest()
ecoli_NP_CP <- do_permanova_com(ord_np_cp, df_np_cp, "ecoli_RF_preds") %>% unnest()
staph_NP_CP <- do_permanova_com(ord_np_cp, df_np_cp, "staph_RF_preds") %>% unnest()
tetw_NP_CP <- do_permanova_com(ord_np_cp, df_np_cp, "tetw_RF_preds") %>% unnest()
meca_NP_CP <- do_permanova_com(ord_np_cp, df_np_cp, "meca_RF_preds") %>% unnest()


# Convert all NAs in the specified columns to 0
df_np_cp$KRD_nGoatsWghtDist.3000m.sum <- replace(df_np_cp$KRD_nGoatsWghtDist.3000m.sum, is.na(df_np_cp$KRD_nGoatsWghtDist.3000m.sum), 0)
df_np_cp$KRD_nPoultryWghtDist.3000m.sum <- replace(df_np_cp$KRD_nPoultryWghtDist.3000m.sum, is.na(df_np_cp$KRD_nPoultryWghtDist.3000m.sum), 0)
df_np_cp$KRD_nPigsWghtDist.3000m.sum <- replace(df_np_cp$KRD_nPigsWghtDist.3000m.sum, is.na(df_np_cp$KRD_nPigsWghtDist.3000m.sum), 0)
df_np_cp$KRD_nCowsWghtDist.3000m.sum <- replace(df_np_cp$KRD_nCowsWghtDist.3000m.sum, is.na(df_np_cp$KRD_nCowsWghtDist.3000m.sum), 0)

# Goats
do_permanova_goat <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nCowsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_goat <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_goat(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}

# Poultry
do_permanova_poultry <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nCowsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_poultry <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_poultry(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}


# Pigs
do_permanova_pig <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nCowsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_pig <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_pig(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}



# Cattle
do_permanova_cattle <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_cattle <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_cattle(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}
# Goat exposure 
goat_NP_CP <- do_permanova_com_goat(ord_np_cp, df_np_cp, "KRD_nGoatsWghtDist.3000m.sum") %>% unnest()
# Poultry exposures
poultry_NP_CP <- do_permanova_com_poultry(ord_np_cp, df_np_cp, "KRD_nPoultryWghtDist.3000m.sum") %>% unnest()
# Pig exposure
pig_NP_CP <- do_permanova_com_pig(ord_np_cp, df_np_cp, "KRD_nPigsWghtDist.3000m.sum") %>% unnest()
# Cattle exposure
cattle_NP_CP <- do_permanova_com_cattle(ord_np_cp, df_np_cp, "KRD_nCowsWghtDist.3000m.sum") %>% unnest()


# Combine results from multivariable models together
livestock_expo_results_np <-rbind(endo_NP_CP, ecoli_NP_CP, staph_NP_CP, tetw_NP_CP, meca_NP_CP, goat_NP_CP, poultry_NP_CP, pig_NP_CP, cattle_NP_CP)

# Ensure `term` is treated as a factor with the correct order
livestock_expo_results_np <- livestock_expo_results_np %>%
  mutate(term = factor(term, levels = c(
    "KRD_nCowsWghtDist.3000m.sum", "KRD_nPoultryWghtDist.3000m.sum", 
    "KRD_nGoatsWghtDist.3000m.sum", "KRD_nPigsWghtDist.3000m.sum", 
    "meca_RF_preds", "tetw_RF_preds", "staph_RF_preds", 
    "ecoli_RF_preds", "EU_ANNUALAVG_PM10_2020"
  ))) %>%
  arrange(term)  # Ensure data is reordered

# BH correction of p-values
livestock_expo_results_np$q.value <- p.adjust(livestock_expo_results_np$p.value, "BH")

# Add q.value label
livestock_expo_results_np$q.value_label <- ifelse(livestock_expo_results_np$q.value < 0.1, "*", NA)

# Rename `Description`
livestock_expo_results_np <- livestock_expo_results_np %>%
  mutate(Description = case_when(
    term == "KRD_nGoatsWghtDist.3000m.sum"  ~ "Goats",
    term == "KRD_nPoultryWghtDist.3000m.sum" ~ "Chickens",
    term == "KRD_nPigsWghtDist.3000m.sum" ~ "Pigs",
    term == "KRD_nCowsWghtDist.3000m.sum" ~ "Cattle",
    term == "EU_ANNUALAVG_PM10_2020" ~ "Endotoxin exposure",
    term == "ecoli_RF_preds" ~ "E. coli exposure",
    term == "staph_RF_preds" ~ "Staph spp. exposure",
    term == "tetw_RF_preds" ~ "tetW exposure",
    term == "meca_RF_preds" ~ "mecA exposure",
    TRUE ~ NA_character_
  ))

# Convert R² values to percentage
livestock_expo_results_np$R2_perc <- livestock_expo_results_np$R2 * 100

# Dot plot
exposure_perm_results_NP <- ggplot(livestock_expo_results_np, aes(x = R2_perc, y = fct_relevel(Description, unique(Description)))) +
  geom_point(aes(fill = q.value, size = R2_perc), shape = 21, alpha = 0.8) + 
  scale_fill_gradient2(
    low = "#1a1143",
    mid = "#95b6b2",
    high = "#e6e6c1",
    limits = c(0.15, 0.988),
    name = "q value"
  ) +
  scale_size_continuous(name = "% Variance Explained", range = c(3, 8)) +
  geom_text(aes(label = q.value_label), size = 7, na.rm = TRUE, color = "red3", vjust = -0.3) +
  theme_minimal() +
  labs(
    x = "% Variance Explained",
    y = "Terms"
  ) +
  theme(
    axis.text.y = element_text(size = 10),
    axis.title = element_text(size = 12),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_NP

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Dot_plot_multivarPERM_livestock_exposures_NP.svg", exposure_perm_results_OP, width= 8, height = 3.5)





# Bar plot of R2
exposure_perm_results_NP <- ggplot(livestock_expo_results_np, 
                                   aes(x = fct_relevel(Description, unique(Description)), 
                                       y = R2_perc)) +
  geom_col(fill = "#dd968e", alpha = 0.8) +  # Single color for bars
  geom_text(aes(label = q.value_label), size = 6, na.rm = TRUE, color = "red3", hjust = -0.2) + 
  coord_flip() +  # Flip axes for better readability
  theme_minimal() +
  labs(
    x = "Terms",
    y = "% Variance Explained"
  ) +
  theme(
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 12),
    axis.title = element_text(size = 14),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Bar_plot_multivarPERM_livestock_exposures_NP.svg", 
       exposure_perm_results_NP, width= 5, height = 4)

```

##### check model assumptions 
```{r}
# Test homogeneity of dispersion

# Goats
# Binning the variable into quartiles
df_np_cp$KRD_nGoatsWghtDist.3000m.sum_binned <- cut(
  df_np_cp$KRD_nGoatsWghtDist.3000m.sum, 
  breaks = quantile(df_np_cp$KRD_nGoatsWghtDist.3000m.sum, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

ord_np_cp <- vegdist(otu_np_cp, method = "bray")  # Keep this as a 'dist' object

dispersion_test <- betadisper(ord_np_cp, df_np_cp$KRD_nGoatsWghtDist.3000m.sum_binned)
anova(dispersion_test)
plot(dispersion_test)

# Cattle
# Binning the variable into quartiles
df_np_cp$KRD_nCowsWghtDist.3000m.sum_binned <- cut(
  df_np_cp$KRD_nCowsWghtDist.3000m.sum, 
  breaks = quantile(df_np_cp$KRD_nCowsWghtDist.3000m.sum, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_cows <- betadisper(ord_np_cp, df_np_cp$KRD_nCowsWghtDist.3000m.sum_binned)
anova(dispersion_test_cows)
plot(dispersion_test_cows)

# Poultry
# Binning the variable into quartiles
df_np_cp$KRD_nPoultryWghtDist.3000m.sum_binned <- cut(
  df_np_cp$KRD_nPoultryWghtDist.3000m.sum, 
  breaks = quantile(df_np_cp$KRD_nPoultryWghtDist.3000m.sum, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_poultry <- betadisper(ord_np_cp, df_np_cp$KRD_nPoultryWghtDist.3000m.sum_binned)
anova(dispersion_test_poultry)
plot(dispersion_test_poultry)

# Pigs
# Binning the variable into quartiles
df_np_cp$KRD_nPigsWghtDist.3000m.sum_binned <- cut(
  df_np_cp$KRD_nPigsWghtDist.3000m.sum, 
  breaks = quantile(df_np_cp$KRD_nPigsWghtDist.3000m.sum, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_pigs <- betadisper(ord_np_cp, df_np_cp$KRD_nPigsWghtDist.3000m.sum_binned)
anova(dispersion_test_pigs)
plot(dispersion_test_pigs)

# E. coli
# Binning the variable into quartiles
df_np_cp$ecoli_RF_preds_binned <- cut(
  df_np_cp$ecoli_RF_preds, 
  breaks = quantile(df_np_cp$ecoli_RF_preds, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_ecoli <- betadisper(ord_np_cp, df_np_cp$ecoli_RF_preds_binned)
anova(dispersion_test_ecoli)
plot(dispersion_test_ecoli)

# Staph
# Binning the variable into quartiles
df_np_cp$staph_RF_preds_binned <- cut(
  df_np_cp$staph_RF_preds, 
  breaks = quantile(df_np_cp$staph_RF_preds, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_staph <- betadisper(ord_np_cp, df_np_cp$staph_RF_preds_binned)
anova(dispersion_test_staph)
plot(dispersion_test_staph)

# tetw
# Binning the variable into quartiles
df_np_cp$tetw_RF_preds_binned <- cut(
  df_np_cp$tetw_RF_preds, 
  breaks = quantile(df_np_cp$tetw_RF_preds, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_tetw <- betadisper(ord_np_cp, df_np_cp$tetw_RF_preds_binned)
anova(dispersion_test_tetw)
plot(dispersion_test_tetw)

# meca
# Binning the variable into quartiles
df_np_cp$meca_RF_preds_binned <- cut(
  df_np_cp$meca_RF_preds, 
  breaks = quantile(df_np_cp$meca_RF_preds, probs = seq(0, 1, 0.25), na.rm = TRUE), 
  include.lowest = TRUE
)

dispersion_test_meca <- betadisper(ord_np_cp, df_np_cp$meca_RF_preds_binned)
anova(dispersion_test_meca)
plot(dispersion_test_meca)


```


#### OP
```{r}
# Microbial exposures
df_op_cp$EU_ANNUALAVG_PM10_2020
endo_op_CP <- do_permanova_com(ord_op_cp, df_op_cp, "EU_ANNUALAVG_PM10_2020") %>% unnest()
ecoli_op_CP <- do_permanova_com(ord_op_cp, df_op_cp, "ecoli_RF_preds") %>% unnest()
staph_op_CP <- do_permanova_com(ord_op_cp, df_op_cp, "staph_RF_preds") %>% unnest()
tetw_op_CP <- do_permanova_com(ord_op_cp, df_op_cp, "tetw_RF_preds") %>% unnest()
meca_op_CP <- do_permanova_com(ord_op_cp, df_op_cp, "meca_RF_preds") %>% unnest()


# Convert all NAs in the specified columns to 0
df_op_cp$KRD_nGoatsWghtDist.3000m.sum <- replace(df_op_cp$KRD_nGoatsWghtDist.3000m.sum, is.na(df_op_cp$KRD_nGoatsWghtDist.3000m.sum), 0)
df_op_cp$KRD_nPoultryWghtDist.3000m.sum <- replace(df_op_cp$KRD_nPoultryWghtDist.3000m.sum, is.na(df_op_cp$KRD_nPoultryWghtDist.3000m.sum), 0)
df_op_cp$KRD_nPigsWghtDist.3000m.sum <- replace(df_op_cp$KRD_nPigsWghtDist.3000m.sum, is.na(df_op_cp$KRD_nPigsWghtDist.3000m.sum), 0)
df_op_cp$KRD_nCowsWghtDist.3000m.sum <- replace(df_op_cp$KRD_nCowsWghtDist.3000m.sum, is.na(df_op_cp$KRD_nCowsWghtDist.3000m.sum), 0)

# Goats
do_permanova_goat <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nCowsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_goat <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_goat(ord = .$ord_sub, 
                                                       meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  # Assign "*" for any p-value < 0.1
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA)
  )
  
  return(t)
}


# Poultry
do_permanova_poultry <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum + KRD_nCowsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_poultry <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_poultry(ord = .$ord_sub, 
                                                          meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA)
  )
  
  return(t)
}


# Pigs
do_permanova_pig <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nCowsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_pig <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_pig(ord = .$ord_sub, 
                                                      meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA)
  )
  
  return(t)
}

# Cattle
do_permanova_cattle <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum + KRD_nGoatsWghtDist.3000m.sum + KRD_nPigsWghtDist.3000m.sum")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_cattle <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_cattle(ord = .$ord_sub, 
                                                         meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA)
  )
  
  return(t)
}
# Goat exposure 
goat_op_CP <- do_permanova_com_goat(ord_op_cp, df_op_cp, "KRD_nGoatsWghtDist.3000m.sum") %>% unnest()
# Poultry exposures
poultry_op_CP <- do_permanova_com_poultry(ord_op_cp, df_op_cp, "KRD_nPoultryWghtDist.3000m.sum") %>% unnest()
# Pig exposure
pig_op_CP <- do_permanova_com_pig(ord_op_cp, df_op_cp, "KRD_nPigsWghtDist.3000m.sum") %>% unnest()
# Cattle exposure
cattle_op_CP <- do_permanova_com_cattle(ord_op_cp, df_op_cp, "KRD_nCowsWghtDist.3000m.sum") %>% unnest()


library(dplyr)
library(forcats)  # For explicit factor reordering

# Combine results from multivariable models together
livestock_expo_results_op <- rbind(endo_op_CP, ecoli_op_CP, staph_op_CP, 
                                   tetw_op_CP, meca_op_CP, goat_op_CP, 
                                   poultry_op_CP, pig_op_CP, cattle_op_CP)

# Ensure `term` is treated as a factor with the correct order
livestock_expo_results_op <- livestock_expo_results_op %>%
  mutate(term = factor(term, levels = c(
    "KRD_nCowsWghtDist.3000m.sum", "KRD_nPoultryWghtDist.3000m.sum", 
    "KRD_nGoatsWghtDist.3000m.sum", "KRD_nPigsWghtDist.3000m.sum", 
    "meca_RF_preds", "tetw_RF_preds", "staph_RF_preds", 
    "ecoli_RF_preds", "EU_ANNUALAVG_PM10_2020"
  ))) %>%
  arrange(term)  # Ensure data is reordered

# BH correction of p-values
livestock_expo_results_op$q.value <- p.adjust(livestock_expo_results_op$p.value, "BH")

# Add q.value label
livestock_expo_results_op$q.value_label <- ifelse(livestock_expo_results_op$q.value < 0.1, "*", NA)

# Rename `Description`
livestock_expo_results_op <- livestock_expo_results_op %>%
  mutate(Description = case_when(
    term == "KRD_nGoatsWghtDist.3000m.sum"  ~ "Goats",
    term == "KRD_nPoultryWghtDist.3000m.sum" ~ "Chickens",
    term == "KRD_nPigsWghtDist.3000m.sum" ~ "Pigs",
    term == "KRD_nCowsWghtDist.3000m.sum" ~ "Cattle",
    term == "EU_ANNUALAVG_PM10_2020" ~ "Endotoxin exposure",
    term == "ecoli_RF_preds" ~ "E. coli exposure",
    term == "staph_RF_preds" ~ "Staph spp. exposure",
    term == "tetw_RF_preds" ~ "tetW exposure",
    term == "meca_RF_preds" ~ "mecA exposure",
    TRUE ~ NA_character_
  ))

# Convert R² values to percentage
livestock_expo_results_op$R2_perc <- livestock_expo_results_op$R2 * 100

# Dot plot
exposure_perm_results_OP <- ggplot(livestock_expo_results_op, aes(x = R2_perc, y = fct_relevel(Description, unique(Description)))) +
  geom_point(aes(fill = q.value, size = R2_perc), shape = 21, alpha = 0.8) + 
  scale_fill_gradient2(
    low = "#1a1143",
    mid = "#95b6b2",
    high = "#e6e6c1",
    limits = c(0.15, 0.988),
    name = "q value"
  ) +
  scale_size_continuous(name = "% Variance Explained", range = c(3, 8)) +
  geom_text(aes(label = q.value_label), size = 7, na.rm = TRUE, color = "red3", vjust = -0.3) +
  theme_minimal() +
  labs(
    x = "% Variance Explained",
    y = "Terms"
  ) +
  theme(
    axis.text.y = element_text(size = 10),
    axis.title = element_text(size = 12),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_OP

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Dot_plot_multivarPERM_livestock_exposures_OP.svg", exposure_perm_results_OP, width= 8, height = 3.5)




# Bar plot of R2
exposure_perm_results_OP <- ggplot(livestock_expo_results_op, 
                                   aes(x = fct_relevel(Description, unique(Description)), 
                                       y = R2_perc)) +
  geom_col(fill = "#547484", alpha = 0.8) +  # Single color for bars
  geom_text(aes(label = q.value_label), size = 6, na.rm = TRUE, color = "red3", hjust = -0.2) + 
  coord_flip() +  # Flip axes for better readability
  theme_minimal() +
  labs(
    x = "Terms",
    y = "% Variance Explained"
  ) +
  theme(
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 12),
    axis.title = element_text(size = 14),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_OP

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Bar_plot_multivarPERM_livestock_exposures_OP.svg", 
       exposure_perm_results_OP, width= 5, height = 4)




```

### Standardised variables
n.b. standardising doesn't alter the R2 of the variables in the PERMANOVAs
#### NP 
```{r}
# Microbial exposures
endo_NP_CP_std <- do_permanova_com(ord_np_cp, df_np_cp, "EU_ANNUALAVG_PM10_2020_std") %>% unnest()
ecoli_NP_CP_std <- do_permanova_com(ord_np_cp, df_np_cp, "ecoli_RF_preds_std") %>% unnest()
staph_NP_CP_std <- do_permanova_com(ord_np_cp, df_np_cp, "staph_RF_preds_std") %>% unnest()
tetw_NP_CP_std <- do_permanova_com(ord_np_cp, df_np_cp, "tetw_RF_preds_std") %>% unnest()
meca_NP_CP_std <- do_permanova_com(ord_np_cp, df_np_cp, "meca_RF_preds_std") %>% unnest()

# Convert all NAs in the specified columns to 0
df_np_cp$KRD_nGoatsWghtDist.3000m.sum_std <- replace(df_np_cp$KRD_nGoatsWghtDist.3000m.sum_std, is.na(df_np_cp$KRD_nGoatsWghtDist.3000m.sum_std), 0)
df_np_cp$KRD_nPoultryWghtDist.3000m.sum_std <- replace(df_np_cp$KRD_nPoultryWghtDist.3000m.sum_std, is.na(df_np_cp$KRD_nPoultryWghtDist.3000m.sum_std), 0)
df_np_cp$KRD_nPigsWghtDist.3000m.sum_std <- replace(df_np_cp$KRD_nPigsWghtDist.3000m.sum_std, is.na(df_np_cp$KRD_nPigsWghtDist.3000m.sum_std), 0)
df_np_cp$KRD_nCowsWghtDist.3000m.sum_std <- replace(df_np_cp$KRD_nCowsWghtDist.3000m.sum_std, is.na(df_np_cp$KRD_nCowsWghtDist.3000m.sum_std), 0)

# Goats
do_permanova_goat_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nCowsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_goat_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_goat_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}

# Poultry
do_permanova_poultry_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nCowsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_poultry_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_poultry_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}


# Pigs
do_permanova_pig_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nCowsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_pig_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_pig_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}



# Cattle
do_permanova_cattle_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_cattle_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_cattle_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}
# Goat exposure 
goat_NP_CP_std <- do_permanova_com_goat_std(ord_np_cp, df_np_cp, "KRD_nGoatsWghtDist.3000m.sum_std") %>% unnest()
# Poultry exposures
poultry_NP_CP_std <- do_permanova_com_poultry_std(ord_np_cp, df_np_cp, "KRD_nPoultryWghtDist.3000m.sum_std") %>% unnest()
# Pig exposure
pig_NP_CP_std <- do_permanova_com_pig_std(ord_np_cp, df_np_cp, "KRD_nPigsWghtDist.3000m.sum_std") %>% unnest()
# Cattle exposure
cattle_NP_CP_std <- do_permanova_com_cattle_std(ord_np_cp, df_np_cp, "KRD_nCowsWghtDist.3000m.sum_std") %>% unnest()


# Combine results from multivariable models together
livestock_expo_results_np_std <-rbind(endo_NP_CP_std, ecoli_NP_CP_std, staph_NP_CP_std, tetw_NP_CP_std, meca_NP_CP_std, goat_NP_CP_std, poultry_NP_CP_std, pig_NP_CP_std, cattle_NP_CP_std)

# Ensure `term` is treated as a factor with the correct order
livestock_expo_results_np_std <- livestock_expo_results_np_std %>%
  mutate(term = factor(term, levels = c(
    "KRD_nCowsWghtDist.3000m.sum_std", "KRD_nPoultryWghtDist.3000m.sum_std", 
    "KRD_nGoatsWghtDist.3000m.sum_std", "KRD_nPigsWghtDist.3000m.sum_std", 
    "meca_RF_preds_std", "tetw_RF_preds_std", "staph_RF_preds_std", 
    "ecoli_RF_preds_std", "EU_ANNUALAVG_PM10_2020_std"
  ))) %>%
  arrange(term)  # Ensure data is reordered

# BH correction of p-values
livestock_expo_results_np_std$q.value <- p.adjust(livestock_expo_results_np_std$p.value, "BH")

# Add q.value label
livestock_expo_results_np_std$q.value_label <- ifelse(livestock_expo_results_np_std$q.value < 0.1, "*", NA)

# Rename `Description`
livestock_expo_results_np_std <- livestock_expo_results_np_std %>%
  mutate(Description = case_when(
    term == "KRD_nGoatsWghtDist.3000m.sum_std"  ~ "Goats",
    term == "KRD_nPoultryWghtDist.3000m.sum_std" ~ "Chickens",
    term == "KRD_nPigsWghtDist.3000m.sum_std" ~ "Pigs",
    term == "KRD_nCowsWghtDist.3000m.sum_std" ~ "Cattle",
    term == "EU_ANNUALAVG_PM10_2020_std" ~ "Endotoxin exposure",
    term == "ecoli_RF_preds_std" ~ "E. coli exposure",
    term == "staph_RF_preds_std" ~ "Staph spp. exposure",
    term == "tetw_RF_preds_std" ~ "tetW exposure",
    term == "meca_RF_preds_std" ~ "mecA exposure",
    TRUE ~ NA_character_
  ))

# Convert R² values to percentage
livestock_expo_results_np_std$R2_perc <- livestock_expo_results_np_std$R2 * 100

# Dot plot
exposure_perm_results_NP_std <- ggplot(livestock_expo_results_np_std, aes(x = R2_perc, y = fct_relevel(Description, unique(Description)))) +
  geom_point(aes(fill = q.value, size = R2_perc), shape = 21, alpha = 0.8) + 
  scale_fill_gradient2(
    low = "#1a1143",
    mid = "#95b6b2",
    high = "#e6e6c1",
    limits = c(0.15, 0.988),
    name = "q value"
  ) +
  scale_size_continuous(name = "% Variance Explained", range = c(3, 8)) +
  geom_text(aes(label = q.value_label), size = 7, na.rm = TRUE, color = "red3", vjust = -0.3) +
  theme_minimal() +
  labs(
    x = "% Variance Explained",
    y = "Terms"
  ) +
  theme(
    axis.text.y = element_text(size = 10),
    axis.title = element_text(size = 12),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_NP_std

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Dot_plot_multivarPERM_livestock_exposures_NP_standardised.svg", exposure_perm_results_NP_std, width= 8, height = 3.5)



# Bar plot of R2
exposure_perm_results_NP_std <- ggplot(livestock_expo_results_np_std, 
                                   aes(x = fct_relevel(Description, unique(Description)), 
                                       y = R2_perc)) +
  geom_col(fill = "#dd968e", alpha = 0.8) +  # Single color for bars
  geom_text(aes(label = q.value_label), size = 6, na.rm = TRUE, color = "red3", hjust = -0.2) + 
  coord_flip() +  # Flip axes for better readability
  theme_minimal() +
  labs(
    x = "Terms",
    y = "% Variance Explained"
  ) +
  theme(
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 12),
    axis.title = element_text(size = 14),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_NP_std
ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Bar_plot_multivarPERM_livestock_exposures_NP_std.svg", 
       exposure_perm_results_NP_std, width= 5, height = 4)

```

#### OP 
```{r}
# Microbial exposures
endo_op_CP_std <- do_permanova_com(ord_op_cp, df_op_cp, "EU_ANNUALAVG_PM10_2020_std") %>% unnest()
ecoli_op_CP_std <- do_permanova_com(ord_op_cp, df_op_cp, "ecoli_RF_preds_std") %>% unnest()
staph_op_CP_std <- do_permanova_com(ord_op_cp, df_op_cp, "staph_RF_preds_std") %>% unnest()
tetw_op_CP_std <- do_permanova_com(ord_op_cp, df_op_cp, "tetw_RF_preds_std") %>% unnest()
meca_op_CP_std <- do_permanova_com(ord_op_cp, df_op_cp, "meca_RF_preds_std") %>% unnest()

# Convert all NAs in the specified columns to 0
df_op_cp$KRD_nGoatsWghtDist.3000m.sum_std <- replace(df_op_cp$KRD_nGoatsWghtDist.3000m.sum_std, is.na(df_op_cp$KRD_nGoatsWghtDist.3000m.sum_std), 0)
df_op_cp$KRD_nPoultryWghtDist.3000m.sum_std <- replace(df_op_cp$KRD_nPoultryWghtDist.3000m.sum_std, is.na(df_op_cp$KRD_nPoultryWghtDist.3000m.sum_std), 0)
df_op_cp$KRD_nPigsWghtDist.3000m.sum_std <- replace(df_op_cp$KRD_nPigsWghtDist.3000m.sum_std, is.na(df_op_cp$KRD_nPigsWghtDist.3000m.sum_std), 0)
df_op_cp$KRD_nCowsWghtDist.3000m.sum_std <- replace(df_op_cp$KRD_nCowsWghtDist.3000m.sum_std, is.na(df_op_cp$KRD_nCowsWghtDist.3000m.sum_std), 0)

# Goats
do_permanova_goat_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nCowsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_goat_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_goat_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}

# Poultry
do_permanova_poultry_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std + KRD_nCowsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_poultry_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_poultry_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}


# Pigs
do_permanova_pig_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nCowsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_pig_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_pig_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}



# Cattle
do_permanova_cattle_std <- function(ord, meta, var) {
  t <- adonis2(as.formula(glue("ord ~ {var} + age + sampling_season + gender + smoked_ever + KRD_nPoultryWghtDist.3000m.sum_std + KRD_nGoatsWghtDist.3000m.sum_std + KRD_nPigsWghtDist.3000m.sum_std")), 
               permutations = 999, data = meta)
  return(t)
}

do_permanova_com_cattle_std <- function(ord, meta, var) {
  t <- beta_df(ord, meta, var) %>% { do_permanova_cattle_std(ord = .$ord_sub, 
                                                    meta = .$m_sub, var) } 
  t <- broom::tidy(t)
  t <- subset(t, term == var)
  
  t <- t %>% mutate(
    label = ifelse(p.value < 0.1, "*", NA))

  return(t)
}
# Goat exposure 
goat_op_CP_std <- do_permanova_com_goat_std(ord_op_cp, df_op_cp, "KRD_nGoatsWghtDist.3000m.sum_std") %>% unnest()
# Poultry exposures
poultry_op_CP_std <- do_permanova_com_poultry_std(ord_op_cp, df_op_cp, "KRD_nPoultryWghtDist.3000m.sum_std") %>% unnest()
# Pig exposure
pig_op_CP_std <- do_permanova_com_pig_std(ord_op_cp, df_op_cp, "KRD_nPigsWghtDist.3000m.sum_std") %>% unnest()
# Cattle exposure
cattle_op_CP_std <- do_permanova_com_cattle_std(ord_op_cp, df_op_cp, "KRD_nCowsWghtDist.3000m.sum_std") %>% unnest()


# Combine results from multivariable models together
livestock_expo_results_op_std <-rbind(endo_op_CP_std, ecoli_op_CP_std, staph_op_CP_std, tetw_op_CP_std, meca_op_CP_std, goat_op_CP_std, poultry_op_CP_std, pig_op_CP_std, cattle_op_CP_std)

# Ensure `term` is treated as a factor with the correct order
livestock_expo_results_op_std <- livestock_expo_results_op_std %>%
  mutate(term = factor(term, levels = c(
    "KRD_nCowsWghtDist.3000m.sum_std", "KRD_nPoultryWghtDist.3000m.sum_std", 
    "KRD_nGoatsWghtDist.3000m.sum_std", "KRD_nPigsWghtDist.3000m.sum_std", 
    "meca_RF_preds_std", "tetw_RF_preds_std", "staph_RF_preds_std", 
    "ecoli_RF_preds_std", "EU_ANNUALAVG_PM10_2020_std"
  ))) %>%
  arrange(term)  # Ensure data is reordered

# BH correction of p-values
livestock_expo_results_op_std$q.value <- p.adjust(livestock_expo_results_op_std$p.value, "BH")

# Add q.value label
livestock_expo_results_op_std$q.value_label <- ifelse(livestock_expo_results_op_std$q.value < 0.1, "*", NA)

# Rename `Description`
livestock_expo_results_op_std <- livestock_expo_results_op_std %>%
  mutate(Description = case_when(
    term == "KRD_nGoatsWghtDist.3000m.sum_std"  ~ "Goats",
    term == "KRD_nPoultryWghtDist.3000m.sum_std" ~ "Chickens",
    term == "KRD_nPigsWghtDist.3000m.sum_std" ~ "Pigs",
    term == "KRD_nCowsWghtDist.3000m.sum_std" ~ "Cattle",
    term == "EU_ANNUALAVG_PM10_2020_std" ~ "Endotoxin exposure",
    term == "ecoli_RF_preds_std" ~ "E. coli exposure",
    term == "staph_RF_preds_std" ~ "Staph spp. exposure",
    term == "tetw_RF_preds_std" ~ "tetW exposure",
    term == "meca_RF_preds_std" ~ "mecA exposure",
    TRUE ~ NA_character_
  ))

# Convert R² values to percentage
livestock_expo_results_op_std$R2_perc <- livestock_expo_results_op_std$R2 * 100

# Dot plot
exposure_perm_results_op_std <- ggplot(livestock_expo_results_op_std, aes(x = R2_perc, y = fct_relevel(Description, unique(Description)))) +
  geom_point(aes(fill = q.value, size = R2_perc), shape = 21, alpha = 0.8) + 
  scale_fill_gradient2(
    low = "#1a1143",
    mid = "#95b6b2",
    high = "#e6e6c1",
    limits = c(0.15, 0.988),
    name = "q value"
  ) +
  scale_size_continuous(name = "% Variance Explained", range = c(3, 8)) +
  geom_text(aes(label = q.value_label), size = 7, na.rm = TRUE, color = "red3", vjust = -0.3) +
  theme_minimal() +
  labs(
    x = "% Variance Explained",
    y = "Terms"
  ) +
  theme(
    axis.text.y = element_text(size = 10),
    axis.title = element_text(size = 12),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_op_std

ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Dot_plot_multivarPERM_livestock_exposures_op_standardised.svg", exposure_perm_results_op_std, width= 8, height = 3.5)



# Bar plot of R2
exposure_perm_results_op_std <- ggplot(livestock_expo_results_op_std, 
                                   aes(x = fct_relevel(Description, unique(Description)), 
                                       y = R2_perc)) +
  geom_col(fill = "#547484", alpha = 0.8) +  # Single color for bars
  geom_text(aes(label = q.value_label), size = 6, na.rm = TRUE, color = "red3", hjust = -0.2) + 
  coord_flip() +  # Flip axes for better readability
  theme_minimal() +
  labs(
    x = "Terms",
    y = "% Variance Explained"
  ) +
  theme(
    axis.text.x = element_text(size = 10),
    axis.text.y = element_text(size = 12),
    axis.title = element_text(size = 14),
    plot.title = element_text(hjust = 0.5, face = "bold", size = 14)
  )

exposure_perm_results_op_std
ggsave("../Output/Beta_diversity/Multivariable_PERMANOVAs/Bar_plot_multivarPERM_livestock_exposures_op_std.svg", 
       exposure_perm_results_op_std, width= 5, height = 4)

```
